% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_features}
\alias{combine_features}
\title{Combine variables}
\usage{
combine_features(x, mode = "mean")
}
\arguments{
\item{x}{\eqn{n_0 \times p_k} matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p_k} is the number of variables inside a group}

\item{mode}{character string "mean" for arithmetic mean  or "pca" for first principal component}
}
\description{
Calculates the mean or the first principal component of a group of variables
}
\examples{
n <- 100; p <- 5
x <- matrix(data=stats::rnorm(n=n*p),nrow=n,ncol=p)
mean <- combine_features(x=x,mode="mean")
comp <- combine_features(x=x,mode="pca")
plot(mean,comp)

}
