% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{construct_matrices}
\alias{construct_matrices}
\title{Construct Matrices}
\usage{
construct_matrices(x, group, type, fuse = "mean")
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{group}{\eqn{p}-dimensional vector of group labels or indices, or list with one slot for each group containing the variable labels or indices}

\item{type}{\eqn{p}-dimensional vector}

\item{fuse}{character string \code{"mean"} for arithmetic mean  or \code{"pca"} for first principal component}
}
\value{
See description.
}
\description{
Constructs matrices with (i) the original data concatenated with the inverted data, (ii) one meta-variable for each group, and (iii) one meta-variable for each group in each type.
}
\examples{
n <- 5
p <- 6
x <- matrix(data=rnorm(n*p),nrow=n,ncol=p)
group <- rep(1:2,each=p/2)
type <- rep(x=1,times=p)
x <- construct_matrices(x=x,group=group,type=type)

}
\seealso{
This function is called by \code{\link{corila}()} and thereby \code{\link{cv.corila}()}.
}
