% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{corila-package}
\alias{corila-package}
\title{Sparse modelling with grouped and correlated features
allowing for privileged information}
\description{
The R package \code{corila} implements
"Sparse modelling with grouped and correlated features
allowing for privileged information" (\emph{Rauschenberger, 2026}).
This is the first implementation of a novel algorithm,
sharing information based on assumed adjacencies and observed correlations,
and allowing not only for primary but also for auxiliary predictors.
It builds upon adaptive lasso regression with the
\link[glmnet:glmnet-package]{glmnet-package}.
}
\details{
Use function \code{\link[=cv.corila]{cv.corila()}} for model fitting.
Type \code{library(corila)} and then \code{?cv.corila} or
\code{help("cv.corila")} to open its help file.

See the vignette for further examples.
Type \code{vignette("corila")} or \code{browseVignettes("corila")}
to open the vignette.

This package also includes the wrapper function \code{\link[=multiridge]{multiridge()}}
for multi-penalty ridge regression with
the \link[multiridge:multiridge-package]{multiridge-package}.
}
\examples{
?cv.corila
?coef.cv.corila
?predict.cv.corila

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
First use \code{\link[=cv.corila]{cv.corila()}} to fit the model,
and then \link[=coef.cv.corila]{coef()} to extract coefficients
or \link[=predict.cv.corila]{predict()} to make predictions.
}
\author{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
}
\keyword{documentation}
