% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crossval}
\alias{crossval}
\title{Cross-validation method}
\usage{
crossval(
  x,
  y,
  family,
  group = NULL,
  type = NULL,
  iter = 5,
  nfolds = 10,
  init.multi = FALSE,
  trial = FALSE,
  method = NULL
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{family}{character string "gaussian", "binomial", "poisson" or "cox"}

\item{group}{XXX}

\item{type}{\eqn{p}-dimensional vector}

\item{iter}{number of cross-validation iterations}

\item{nfolds}{number of cross-validation folds}

\item{init.multi}{logical}

\item{trial}{logical}

\item{method}{character vector listing all methods to be compared}
}
\description{
Compares methods with cross-validation method
}
\details{
This function implements repeated $k$-fold cross-validation (e.g., 5 repetitions of 10-fold cross-validation).
}
\examples{
NULL

}
