% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{deviance.cv.corila}
\alias{deviance.cv.corila}
\title{Deviance}
\usage{
\method{deviance}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{cv.corila}}
}
\description{
Calculates the deviance.
}
\details{
Returns the deviance calculated by \code{\link[glmnet:deviance.glmnet]{glmnet::deviance.glmnet()}}
for the model with the optimised mixing and regularisation hyperparameters.
}
\seealso{
The internal function \code{\link[=.deviance]{.deviance()}} calculates
the deviance from fitted and observed values.
}
