% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.assert}
\alias{.assert}
\title{Assertions}
\usage{
.assert(
  x,
  type = "numeric",
  dim = 1L,
  na.rm = FALSE,
  support = NULL,
  family = NULL,
  min = -Inf,
  max = Inf
)
}
\arguments{
\item{x}{scalar, vector (of length \code{dim}),
matrix (of dimensions \code{dim}),
or array (of dimensions \code{dim}) to be checked
\itemize{
\item \code{type = "numeric"}: numeric
\item \code{type = "integer"}: integer
\item \code{type = "nominal"}: character
\item \code{type = "logical"}: logical
\item \code{family = "binomial"}: integers 0 or 1
\item \code{family = "poisson"}: non-negative integers
\item \code{family = "cox"}: object created with \link[survival:Surv]{survival::Surv}
}}

\item{type}{character scalar \code{"numeric"} (default), \code{"integer"},
\code{"nominal"}, or \code{"logical"}}

\item{dim}{vector of length 1, 2 or 3 containing positive integers
(minimum 1, maximum \eqn{100,000})
defining the dimensionality:
\itemize{
\item scalar \code{x}: \code{dim = 1}
\item vector \code{x} of length 100: \code{dim = 100}
\item vector \code{x} of arbitrary length: \code{dim = Inf}
\item matrix \code{x} with 100 rows: \code{dim = c(100, Inf)}
\item matrix \code{x} of arbitrary dimensions: \code{dim = c(Inf, Inf)}
\item array \code{x} of arbitrary dimensions: \code{dim = c(Inf, Inf, Inf)}
}}

\item{na.rm}{logical scalar (or numeric 0/1):
\itemize{
\item \code{na.rm=FALSE} (or \code{na.rm=0}): missing values are not allowed
\item \code{na.rm=TRUE} (or \code{na.rm=1}): missing values are allowed
}}

\item{support}{character vector (only used for \code{type = "nominal"}),
(matching with \code{x} is case-insensitive)}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{min}{numeric scalar (not used for \code{type = "nominal"})}

\item{max}{numeric scalar (not used for \code{type = "nominal"})}
}
\value{
Returns \code{NULL} invisibly, or throws an error.
}
\description{
Check whether provided arguments satisfy expectations.
}
\details{
This function is called by multiple function of the \link{corila-package}.
}
\examples{
\dontshow{.assert <- corila:::.assert}
n <- 3L; p <- 4L
.assert(x = matrix(rnorm(n = n * p), nrow = n, ncol = p),
        dim = c(n, p),
        type = "numeric",
        family = "gaussian")
.assert(x = rpois(n = n, lambda = 4.0),
        dim = n,
        type = "integer",
        family = "poisson")
.assert(x = rbinom(n = n, size = 1L, prob = 0.5),
        dim = n,
        type = "integer",
        family = "binomial")
.assert(x = "a",
        dim = 1L,
        type = "nominal",
        support = letters)

}
\seealso{
The function \code{\link[=.validate]{.validate()}} verifies whether the main arguments
have compatible dimensions (number of samples and features).
}
\keyword{internal}
