% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.deviance}
\alias{.deviance}
\title{Deviance}
\usage{
.deviance(y, y_hat, family)
}
\arguments{
\item{y}{response:
numeric vector of length \code{n}}

\item{y_hat}{predicted response:
numeric vector of length \code{n}}

\item{family}{character}
}
\description{
Calculate the deviance
}
\examples{
n <- 10

y <- rnorm(n)
y_hat <- rnorm(n)
.deviance(y = y , y_hat = y_hat, family = "gaussian")

y <- rbinom(n = n, size = 1, prob = 0.5)
y_hat <- runif(n)
.deviance(y = y , y_hat = y_hat, family = "binomial")

y <- rpois(n = n, lambda = 4)
y_hat <- rexp(n)
.deviance(y = y , y_hat = y_hat, family = "poisson")

}
\keyword{internal}
