% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.is_adjacent}
\alias{.is_adjacent}
\title{Adjacency}
\usage{
.is_adjacent(group, j, p, names)
}
\arguments{
\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\emph{(ii)} list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
or \emph{(iii)} \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{j}{index of predictor}

\item{p}{number of predictors}

\item{names}{names of predictors}
}
\value{
Returns a logical vector of length \eqn{p}.
}
\description{
Identifies adjacent predictors.
}
\examples{
NA

}
\keyword{internal}
