% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.residuals}
\alias{.residuals}
\title{Deviance Residuals}
\usage{
.residuals(y_obs, y_fit, family)
}
\arguments{
\item{y_obs}{\eqn{n}-dimensional vector of observed values}

\item{y_fit}{\eqn{n}-dimensional vector of fitted values or probabilities}

\item{family}{character \code{"gaussian"}, \code{"binomial"}, or \code{"poisson"}}
}
\value{
Returns an \eqn{n}-dimensional vector.
}
\description{
Calculates the deviance residuals.
}
\details{
This function is called by \code{\link[=residuals.cv.corila]{residuals.cv.corila()}}.
}
\examples{
\dontshow{.residuals <- corila:::.residuals}
n <- 10L

y_obs <- stats::rnorm(n = n)
y_fit <- stats::rnorm(n = n)
.residuals(y_obs = y_obs, y_fit = y_fit, family = "gaussian")

y_obs <- stats::rbinom(n = n, size = 1L, prob = 0.2)
y_fit <- stats::runif(n = n)
.residuals(y_obs = y_obs, y_fit = y_fit, family = "binomial")

y_obs <- stats::rpois(n = n, lambda = 4.0)
y_fit <- stats::rexp(n = n, rate = 0.25)
.residuals(y_obs = y_obs, y_fit = y_fit, family = "poisson")

}
\keyword{internal}
