% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.set_candidates}
\alias{.set_candidates}
\title{Candidate Values}
\usage{
.set_candidates(tune)
}
\arguments{
\item{tune}{character string for determining the candidate values
for the hyperparameters
(to implement: list with slots
\code{wgt_local}, \code{wgt_global}, \code{exp_local}, and \code{exp_global})}
}
\value{
Returns a data frame with
the slots \code{"wgt_local"} and \code{"exp_local"}
for the local prior information
and the slots \code{"wgt_global"} and \code{"exp_global"}
for the global prior information.
}
\description{
Set candidate values for hyperparameters.
}
\examples{
corila:::.set_candidates(tune = "none")

}
\keyword{internal}
