% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_effects}
\alias{.simulate_effects}
\title{Simulate effects}
\usage{
.simulate_effects(
  group,
  prob_group = 0.5,
  prob_predictor = 0.8,
  signal_strength = 1,
  seed = 1L
)
}
\arguments{
\item{group}{group indicator:
integer vector of length \eqn{p} with entries between 1 and \eqn{q},
where \eqn{p} is the number of predictors
and \eqn{q} is the number of predictor groups
(maximum length \eqn{1\,000},
minimum entry 1, maximum entry \eqn{1\,000})}

\item{prob_group}{probability for each predictor group to be active:
numeric scalar in the unit interval
(minimum 0 makes all groups inactive,
maximum 1 makes all groups active)}

\item{prob_predictor}{probability for each predictor in an active group to be active:
numeric scalar in the unit interval
(minimum 0 makes all predictors inactive,
maximum 1 makes all predictors in active groups active)}

\item{signal_strength}{non-negative numeric scalar
for multiplying the effect sizes
(default: \code{signal_strength=1.0},
minimum 0 sets all effect sizes to 0,
maximum 2 to avoid undefined values)}

\item{seed}{random seed for reproducibility:
integer scalar (unrestricted)}
}
\value{
Returns a numeric vector of length \eqn{p}.
}
\description{
Simulates effect vector.
}
\examples{
\dontshow{.simulate_effects <- corila:::.simulate_effects}
group <- rep(c(1L:5L), each = 3L)
.simulate_effects(group = group)
.simulate_effects(group = group, signal_strength = 1.5)

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
