% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_outcome}
\alias{.simulate_outcome}
\title{Simulate outcome}
\usage{
.simulate_outcome(family, x = NULL, beta = NULL, n = NULL, factor = 1)
}
\arguments{
\item{family}{character \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{x}{numeric \eqn{n \times p} matrix}

\item{beta}{numeric \eqn{p}-dimensional vector}

\item{n}{integer or \code{NULL}}

\item{factor}{non-negative scalar (default: \code{factor=1})
for multiplying the linear predictor
(to increase or decrease the signal strength)}
}
\value{
Returns an \eqn{n}-dimensional outcome vector.
}
\description{
Simulates outcome vector.
}
\examples{
# simulate independent outcome
corila:::.simulate_outcome(family = "gaussian", n = 10)

# simulate dependent outcome
n <- 10
p <- 20
x <- matrix(rnorm(n * p), n, p)
beta <- rnorm(p)
corila:::.simulate_outcome(family = "gaussian", x = x, beta = beta)

}
\seealso{
Use \code{\link[=simulate]{simulate()}} to simulate a predictor matrix,
an effect vector, and an outcome vector.
}
\keyword{internal}
