% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_predictors}
\alias{.simulate_predictors}
\title{Simulate predictors}
\usage{
.simulate_predictors(n, p = NULL, group = NULL, rho = 0, seed = 1L)
}
\arguments{
\item{n}{number of observations:
positive integer
(minimum 1, maximum \eqn{1\,000\,000})}

\item{p}{number of predictors:
positive integer scalar
(minimum 1 leads to a single predictor,
maximum \eqn{1\,000\,000})}

\item{group}{group indicator:
integer vector of length \eqn{p} with entries between 1 and \eqn{q},
where \eqn{p} is the number of predictors
and \eqn{q} is the number of predictor groups
(maximum length \eqn{1\,000\,000},
minimum entry 1, maximum entry \eqn{1\,000\,000})}

\item{rho}{correlation coefficient for predictors within the same group:
numeric scalar in the unit interval
(minimum 0 leads to uncorrelated predictors within each group,
maximum 1 leads to identical predictors within each group)}

\item{seed}{random seed for reproducibility:
integer scalar (unrestricted)}
}
\value{
Returns a numeric matrix with \eqn{n} rows (observations)
and \eqn{p} columns (predictors).
}
\description{
Simulates predictor matrix.
}
\examples{
\dontshow{.simulate_predictors <- corila:::.simulate_predictors}
.simulate_predictors(n = 5L, p = 7L)
.simulate_predictors(n = 5L, group = rep(c(1L, 2L), each = 3L), rho = 1.0)

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
