% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_response}
\alias{.simulate_response}
\title{Simulate outcome}
\usage{
.simulate_response(family, x = NULL, beta = NULL, n = NULL)
}
\arguments{
\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{x}{predictors:
numeric matrix with \eqn{n} rows (observations)
and \eqn{p} columns (predictors)}

\item{beta}{effects:
numeric vector of length \eqn{q}}

\item{n}{sample size:
positive integer scalar or \code{NULL}}
}
\value{
Returns an \eqn{n}-dimensional response vector.
}
\description{
Simulates outcome vector.
}
\examples{
\dontshow{.simulate_response <- corila:::.simulate_response}
# simulate independent response
.simulate_response(family = "gaussian", n = 10L)

# simulate dependent response
n <- 10L
p <- 20L
x <- matrix(rnorm(n * p), n, p)
beta <- rnorm(p)
.simulate_response(family = "gaussian", x = x, beta = beta)

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
