% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.type}
\alias{.type}
\title{Name method (helper function)}
\usage{
.type(alpha)
}
\arguments{
\item{alpha}{elastic net mixing parameter or character string
(see \code{alpha_init} and \code{alpha_final}
in \code{\link[=cv.corila]{cv.corila()}})}
}
\value{
Returns a character string
(\code{"ridge regression"}, \code{"lasso regression"}, \code{"elastic net regression"},
\code{"multi-penalty ridge regression"},
or \code{"Pearson/Spearman/Kendall correlation"})
}
\description{
Names the method used for obtaining initial or final coefficients.
}
\examples{
corila:::.type(alpha = 0)

}
\seealso{
This function is called by \code{\link[=print.summary.cv.corila]{print.summary.cv.corila()}}.
}
\keyword{internal}
