% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate}
\alias{.validate}
\title{Validate data}
\usage{
.validate(x, y, group, family)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector,
where \eqn{n_0} is the number of observations used for model training}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\emph{(ii)} list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
or \emph{(iii)} \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}
}
\value{
Returns \code{NULL} or an error message.
}
\description{
Validates the predictor matrix x and the outcome vector y.
}
\examples{
n <- 10
p <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
corila:::.validate(x = x, y = y, group = NULL, family = "gaussian")

}
\seealso{
\code{\link{.check}()}
}
\keyword{internal}
