% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand_auxiliary}
\alias{expand_auxiliary}
\title{Expand auxiliary features}
\usage{
expand_auxiliary(x, include)
}
\arguments{
\item{include}{logical vector of length \code{p}}

\item{newx}{matrix with \code{n} rows}
}
\value{
matrix with \code{n} rows and \code{p} columns
}
\description{
Add empty columns for auxiliary features.
}
\examples{
n <- 5
p <- 10
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
include <- as.logical(stats::rbinom(n = p, size = 1, prob = 0.5))
x_primary <- x[,include]
x_expanded <- expand_auxiliary(x = x_primary, include = include)
all(x_expanded[,include]==x[,include]) # move to unit tests
all(x_expanded[,!include]==0) # move to unit tests

}
