% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.cv.corila}
\alias{fitted.cv.corila}
\title{Fitted values}
\usage{
\method{fitted}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(for compatibility with \link[stats:fitted]{stats::fitted})}
}
\value{
Returns a numeric vector of length \eqn{n_0}
(one entry for each training observation).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts fitted values.
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# simulating data
n <- 10L; p <- 20L; q <- 5L
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1L, prob = 0.5))

# fitting the model
object <- cv.corila(x = x, y = y, group = group, primary = primary)

# using S3 methods
coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
\seealso{
Use \link[=predict.cv.corila]{predict()} to obtain predicted values
(i.e., for testing observations).
}
