% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.forescale}
\alias{.forescale}
\title{Standardisation}
\usage{
.forescale(x, y = NULL, family = NULL, pars = NULL)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector
(only required if \code{family="gaussian"})
or \code{NULL}}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"};
or \code{NULL} (if \code{pars} is provided)}

\item{pars}{list as defined in section \emph{Value},
or \code{NULL} (if \code{family} is provided)}
}
\value{
Returns a list with multiple slots:
\itemize{
\item standardised \eqn{n_0 \times p} or \eqn{n_1 \times p}
predictor matrix \eqn{x}
\item standardised \eqn{n_0}-dimensional or \eqn{n_1}-dimensional
response vector \eqn{y}
(only if \eqn{y} is provided and \code{family = "gaussian"}
or \code{pars$family = "gaussian"}; otherwise output equals input)
\item character string \code{family} indicates the model (\code{"gaussian"},
\code{"binomial"}, \code{"poisson"}, or \code{"cox"}),
determined by argument \code{family} or \code{pars$family}
\item list \code{pars} with slots \code{mu.x} and \code{sd.x}
(\eqn{p}-dimensional vectors of means and standard deviations
of the predictor variables),
\code{mu.y} and \code{sd.y}
(mean and standard deviation of response variable for Gaussian family,
0 and 1 for other families),
and \code{family}
(character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"})
}
}
\description{
Transforms variables to mean 0 and variance 1.
}
\details{
This function is called by \code{\link[=corila]{corila()}} for the training data
and by \code{\link[=predict.corila]{predict.corila()}} for the testing data.
}
\examples{
\donttest{
\dontshow{
.forescale <- corila:::.forescale
.backscale <- corila:::.backscale
}

# simulate data
family <- "gaussian"
data <- simulate_data(family = family, prob_primary = 1.0)

# regression without standardisation
if (identical(family, "cox")) {
  lm1 <- survival::coxph(data$y_train~., data = data.frame(data$x_train))
} else {
  lm1 <- stats::glm(data$y_train~., data = data.frame(data$x_train),
                    family = family)
}
coef1 <- stats::coef(lm1)
yhat1 <- predict(lm1, newdata = data.frame(data$x_test))

# regression with standardisation
scale <- .forescale(x = data$x_train,
                    y = data$y_train,
                    family = family)
if (identical(family, "cox")) {
  lm2 <- survival::coxph(scale$y~., data = data.frame(scale$x))
} else {
  lm2 <- stats::glm(scale$y~., data = data.frame(scale$x), family = family)
}
coef_temp <- stats::coef(lm2)
newx_temp <- .forescale(x = data$x_test,
                        pars = scale$pars)$x
yhat_temp <- predict(object = lm2, newdata = data.frame(newx_temp))
result <- .backscale(pars = scale$pars,
                     y = yhat_temp,
                     coef = coef_temp)
coef2 <- result$coef
yhat2 <- result$y

# equality
all.equal(coef1, coef2)
all.equal(yhat1, yhat2)
\dontshow{
stopifnot(isTRUE(all.equal(coef1, coef2)), isTRUE(all.equal(yhat1, yhat2)))
}
}

}
\seealso{
Use function \code{\link[=.backscale]{.backscale()}}
to bring coefficients and predictions back to original scale.
}
\keyword{internal}
