% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{holdout}
\alias{holdout}
\title{Comparison with hold-out}
\usage{
holdout(
  x_train,
  y_train,
  group,
  type,
  family,
  x_test = NULL,
  y_test = NULL,
  nfolds = 10,
  foldid = NULL,
  method = NULL,
  seed = NULL,
  init.multi = FALSE,
  trial = TRUE
)
}
\arguments{
\item{x_train}{\eqn{n_0 \times p} matrix}

\item{y_train}{\eqn{n_0}-dimensional vector}

\item{group}{XXX}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string "gaussian", "binomial", "poisson" or "cox"}

\item{x_test}{\eqn{n_1 \times p} matrix}

\item{y_test}{\eqn{n_1}-dimensional vector}

\item{nfolds}{number of internal cross-validation folds (integer scalar)}

\item{foldid}{internal cross-validation fold identifiers (\eqn{p}-dimensional integer vector)}

\item{method}{character vector listing all methods to be compared}

\item{seed}{random seed (integer scalar)}

\item{init.multi}{logical}

\item{trial}{logical}
}
\description{
Compares methods using hold-out method
}
\examples{
NULL

}
