% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.is_adjacent}
\alias{.is_adjacent}
\title{Adjacency indicator}
\usage{
.is_adjacent(group, j, p, names)
}
\arguments{
\item{group}{group structure (multiple options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
(usually with a unit diagonal so that predictors retain the information
from their initial coefficient)
}}

\item{j}{index of predictor
(positive integer between \eqn{1} and \eqn{p})}

\item{p}{number of predictors
(positive integer)}

\item{names}{names of predictors
(character vector of length \eqn{p})}
}
\value{
Returns a logical vector of length \eqn{p}.
}
\description{
Identifies adjacent predictors.
}
\details{
This function is called by \code{\link[=corila]{corila()}}.
A predictor is adjacent to itself.
If argument \code{group} is a list (specifying potentially overlapping groups),
two predictors are adjacent if they are one or more common groups.
}
\examples{
\dontshow{.is_adjacent <- corila:::.is_adjacent}
p <- 5L
names <- paste0("x", seq_len(p))
group <- list()
group$index_vector <- setNames(object = c(1L, 1L, 2L, 2L, 3L), nm = names)
group$label_vector <- setNames(object = LETTERS[group$index_vector],
                                 nm = names(group$index_vector))
group$index_list <- lapply(X = setNames(nm = unique(group$label_vector)),
                     FUN = function(x) which(group$label_vector == x))
group$label_list <- lapply(group$index_list, names)
group$matrix <- 1L * outer(X = group$index_vector,
                          Y = group$index_vector,
                          FUN = "==")
.is_adjacent(group = group[[1L]], j = 3L, p = p, names = names)

}
\keyword{internal}
