% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{nobs.cv.corila}
\alias{nobs.cv.corila}
\title{Observation Count}
\usage{
\method{nobs}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{cv.corila}}
}
\description{
Extracts the number of observations.
}
