% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.cv.corila}
\alias{plot.cv.corila}
\title{Plot Sparse Group Lasso (S3 method)}
\usage{
\method{plot}{cv.corila}(x, ...)
}
\arguments{
\item{...}{(not used)}

\item{object}{object of class \code{"cv.corila"}}
}
\value{
Returns NULL (invisible).
}
\description{
Plot method for class \code{"cv.corila"}.
}
