% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cv.corila}
\alias{print.cv.corila}
\title{print (S3 method)}
\usage{
\method{print}{cv.corila}(x, ...)
}
\arguments{
\item{x}{object of class \code{"cv.corila"}}

\item{...}{(for compatibility with \link[base:print]{base::print})}
}
\value{
Prints \code{"object of class 'cv.corila'"} to the console
(with a note on the number of cross-validated models).
Returns \code{x} invisibly.
}
\description{
Print method for class \code{"cv.corila"}.
}
\examples{
n <- 12L # decrease to 10 to check LOOCV
p <- 20L
q <- 5L
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1L, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)
print(object)
summary(object)

}
\seealso{
\link[=summary.cv.corila]{summary()}
}
\keyword{internal}
