% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{residuals.cv.corila}
\alias{residuals.cv.corila}
\title{Residuals}
\usage{
\method{residuals}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns a numeric vector of length \eqn{n_0}
(one entry for each training observation).
}
\description{
Residuals
}
\details{
This function extracts the observed and fitted values from the fitted model
and calls the internal function \code{\link[=.residuals]{.residuals()}} to calculate the residuals.
}
