% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.residuals}
\alias{.residuals}
\title{Deviance Residuals}
\usage{
.residuals(y, y_hat, family)
}
\arguments{
\item{y}{\eqn{n_0}-dimensional vector of observed values}

\item{y_hat}{\eqn{n_0}-dimensional vector of fitted values or probabilities}

\item{family}{character \code{"gaussian"}, \code{"binomial"}, or \code{"poisson"}}
}
\value{
Returns an \eqn{n_0}-dimensional vector.
}
\description{
Calculates the deviance residuals.
}
\details{
This function is called by \code{\link[=residuals.cv.corila]{residuals.cv.corila()}}.
}
\examples{
\dontshow{.residuals <- corila:::.residuals}
n <- 10L

y <- stats::rnorm(n = n)
y_hat <- stats::rnorm(n = n)
.residuals(y = y, y_hat = y_hat, family = "gaussian")

y <- stats::rbinom(n = n, size = 1L, prob = 0.2)
y_hat <- stats::runif(n = n)
.residuals(y = y, y_hat = y_hat, family = "binomial")

y <- stats::rpois(n = n, lambda = 4.0)
y_hat <- stats::rexp(n = n, rate = 0.25)
.residuals(y = y, y_hat = y_hat, family = "poisson")

}
\keyword{internal}
