% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.set_candidates}
\alias{.set_candidates}
\title{Candidate values}
\usage{
.set_candidates(tune)
}
\arguments{
\item{tune}{character string for determining the candidate values
for the hyperparameters:
\itemize{
\item \code{"none"}:
fixed weights and exponents
(\code{wgt_local}=1, \code{exp_local}=0, \code{wgt_global}=0),
no tuning
(share information within groups, ignore correlations)
\item \code{"weight"} (default):
fixed exponents (\code{exp_local}=0, \code{exp_global}=2),
tuning \code{wgt_local}=1-\code{wgt_global}
(find a compromise between sharing information within groups
and between correlated predictors)
\item \code{"exponent"}:
fixed weights (\code{wgt_local}=1, \code{wgt_global}=0),
tuning \code{exp_local}
(share information between correlated predictors in the same group,
determine level of trust in correlation coefficients)
\item \code{"bivariate"}:
tuning \code{wgt_local}=1-\code{wgt_global} and \code{exp_local}=\code{exp_global}
(find a compromise between sharing information
between predictors in the same group and between all predictors,
determine level of trust in correlation coefficients)
\item \code{"factorial"}:
tuning \code{wgt_local}=1-\code{wgt_global}, \code{exp_local}, and \code{exp_global}
(find a compromise between sharing information within groups
and between correlated predictors,
with possibly different exponents)
}

(The internal function \code{\link[=.set_candidates]{.set_candidates()}} uses this argument
to create a grid of candidates values for
\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global}.)}
}
\value{
Returns a data frame with
the slots \code{"wgt_local"} and \code{"exp_local"}
for the local prior information
and the slots \code{"wgt_global"} and \code{"exp_global"}
for the global prior information.
}
\description{
Sets candidate values for hyperparameters.
}
\details{
\itemize{
\item If the local weight equals 0, the local exponent has no influence.
And if the global weight equals 0, the global exponent has no influence.
Therefore, if a weight is close to zero, its exponent is set to infinity.
This avoids redundant combinations of hyperparameters
(i.e., a local weight of zero with multiple local exponents,
or a global weight of zero with multiple global exponents)
\item The experimental hyperparameter \code{"threshold"} is currently always set to 0
(no thresholding of correlation coefficients).
}
}
\examples{
\dontshow{.set_candidates <- corila:::.set_candidates}
.set_candidates(tune = "none")

}
\seealso{
This function is called by \code{\link[=cv.corila]{cv.corila()}}.
}
\keyword{internal}
