% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Data simulation}
\usage{
simulate(
  family = "gaussian",
  n0 = 100,
  n1 = 10000,
  n.group = 20,
  n.type = 2,
  size.group = c(5, 3),
  effect.size = c(1, 1),
  corfac.feature = 0.5,
  corfac.type = 0.5,
  corfac.group = 0.25,
  n.group.causal = 2,
  prop.causal = 0.5,
  noise.factor = 1,
  plot = TRUE
)
}
\arguments{
\item{family}{character "gaussian", "binomial", "poisson" or "cox"}

\item{n0}{number of training observations}

\item{n1}{number of testing observations}

\item{n.group}{number of variable groups}

\item{n.type}{number of variable types}

\item{size.group}{size of variable groups (per variable type)}

\item{effect.size}{effect sizes (per variable type)}

\item{corfac.feature}{decrease of correlation if different variable}

\item{corfac.type}{decrease of correlation if different type}

\item{corfac.group}{decrease of correlation if different group}

\item{n.group.causal}{number of causal groups}

\item{prop.causal}{proportion of causal features within causal groups}

\item{noise.factor}{noise factor}

\item{plot}{logical}
}
\description{
Simulates data with grouped predictor variables
}
\examples{
NULL

}
