% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\title{Data simulation}
\usage{
simulate_data(
  n0 = 50L,
  n1 = 20L,
  p = 30L,
  q = 10L,
  family = "gaussian",
  rho_between = 0,
  rho_within = 0.5,
  prob_primary = 0.5,
  prob_group = 0.5,
  prob_predictor = 0.8,
  signal_strength = 1,
  seed = 1L
)
}
\arguments{
\item{n0}{number of training observations:
positive integer scalar
(minimum 1,
maximum \eqn{10\,000})}

\item{n1}{number of testing observations:
non-negative integer scalar
(minimum 0,
maximum \eqn{100\,000})}

\item{p}{number of predictors:
positive integer scalar
(minimum 1 leads to a single predictor,
maximum \eqn{1\,000})}

\item{q}{number of predictor groups:
positive integer scalar
(minimum 1 assigns all predictors to the same group.
maximum \code{p} assigns each predictor to its own group)}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{rho_between}{correlation coefficient for predictors in different groups:
positive numeric scalar in the unit interval
(minimum 0 leads to uncorrelated predictors between groups,
maximum \code{rho_within} leads to same correlation between and within groups)}

\item{rho_within}{correlation coefficient for predictors within the same group:
positive numeric scalar in the unit interval
(minimum 0 leads to uncorrelated predictors within each group,
maximum 1 leads to identical predictors within each group)}

\item{prob_primary}{probability for each predictor to be primary (rather than auxiliary):
numeric scalar in the unit interval
(minimum 0 leads to auxiliary predictors only,
maximum 1 leads to primary predictors only)}

\item{prob_group}{probability for each predictor group to be active:
numeric scalar in the unit interval
(minimum 0 makes all groups inactive,
maximum 1 makes all groups active)}

\item{prob_predictor}{probability for each predictor in an active group to be active:
numeric scalar in the unit interval
(minimum 0 makes all predictors inactive,
maximum 1 makes all predictors in active groups active)}

\item{signal_strength}{non-negative numeric scalar
for multiplying the effect sizes
(default: \code{signal_strength=1.0},
minimum 0 sets all effect sizes to 0,
maximum 2 to avoid undefined values)}

\item{seed}{random seed for reproducibility:
integer scalar (unrestricted)}
}
\value{
Returns a named list with the following slots:
\itemize{
\item \code{x_train}:
predictor matrix of the training observations
(\eqn{n_0} rows, \eqn{p} columns)
\item \code{y_train}:
response vector of the training observations
(length \eqn{n_0})
\item \code{group}:
integer vector indicating the group of the predictors
(length \eqn{p})
\item \code{primary}:
logical vector indicating
primary (\code{TRUE}) and auxiliary (\code{FALSE}) predictors
(length \eqn{p})
\item \code{beta}:
numeric vector of the effects of the predictors on the response
(length \eqn{p})
\item \code{x_test}:
\eqn{n_1 \times p} predictor matrix for the test observations
\item \code{y_test}:
response vector for the test observations of length \eqn{n_1}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Simulates a predictor matrix, an effect vector and a response vector.
The simulated datasets can be used for modelling a response based on
grouped and correlated primary and auxiliary predictors.
}
\details{
Use the objects \code{x_train}, \code{y_train}, \code{group}, and \code{primary}
for model training.
Estimated coefficients can be compared with \code{beta}.

Use the object \code{x_test} for model testing.
Predicted values can be compared with \code{y_test}.

Training and testing observations are named \code{train_} or \code{test_},
respectively, followed by a number indexing the observations
(e.g., \code{train_1} or \code{test_1}).

Primary and auxiliary predictors are named \code{pri_} or \code{aux_}, respectively,
followed by a number indexing the predictor groups, a point,
and a number indexing the predictors within this group
(e.g., \code{pri_1.1} or \code{aux_1.1}).
}
\examples{
data <- simulate_data(n0 = 50L, n1 = 20L, p = 30L, q = 10L,
                     family = "gaussian", rho_between = 0.0,
                     rho_within = 0.5, prob_primary = 0.5,
                     signal_strength = 1.0, prob_group = 0.5,
                     prob_predictor = 0.8, seed = 1L)
utils::str(data, vec.len = 2L)

}
\seealso{
This function calls the internal functions \code{\link[=.simulate_predictors]{.simulate_predictors()}},
\code{\link[=.simulate_effects]{.simulate_effects()}}, and \code{\link[=.simulate_response]{.simulate_response()}} for simulating the
predictor matrix, the effect vector, or the response vector, respectively.
}
