% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_predictors}
\alias{.simulate_predictors}
\title{Simulate predictors}
\usage{
.simulate_predictors(
  n,
  p = NULL,
  group = NULL,
  rho_within = 0,
  rho_between = 0,
  seed = 1L
)
}
\arguments{
\item{n}{number of observations:
positive integer
(minimum 1, maximum \eqn{110\,000})}

\item{p}{number of predictors:
positive integer scalar
(minimum 1 leads to a single predictor,
maximum \eqn{1\,000})}

\item{group}{group indicator:
integer vector of length \eqn{p} with entries between 1 and \eqn{q},
where \eqn{p} is the number of predictors
and \eqn{q} is the number of predictor groups
(maximum length \eqn{1\,000},
minimum entry 1, maximum entry \eqn{1\,000})}

\item{rho_within}{correlation coefficient for predictors within the same group:
positive numeric scalar in the unit interval
(minimum 0 leads to uncorrelated predictors within each group,
maximum 1 leads to identical predictors within each group)}

\item{rho_between}{correlation coefficient for predictors in different groups:
positive numeric scalar in the unit interval
(minimum 0 leads to uncorrelated predictors between groups,
maximum \code{rho_within} leads to same correlation between and within groups)}

\item{seed}{random seed for reproducibility:
integer scalar (unrestricted)}
}
\value{
Returns a numeric matrix with \eqn{n} rows (observations)
and \eqn{p} columns (predictors).
}
\description{
Simulates predictor matrix.
}
\examples{
\dontshow{.simulate_predictors <- corila:::.simulate_predictors}
.simulate_predictors(n = 5L, p = 7L)
.simulate_predictors(n = 5L, group = rep(c(1L, 2L), each = 3L),
                     rho_within = 0.5, rho_between = 0.2)

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
