% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.cv.corila}
\alias{summary.cv.corila}
\alias{print.summary.cv.corila}
\title{Summarising Sparse Group Lasso (S3 method)}
\usage{
\method{summary}{cv.corila}(object, ...)

\method{print}{summary.cv.corila}(x, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}

\item{x}{object of class \code{"summary.cv.corila"}}
}
\value{
The function \code{summary.cv.corila} returns
an invisible list with multiple slots.
}
\description{
Summary method for class \code{"cv.corila"}.
}
\seealso{
print.corila
}
