% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{backscale}
\alias{backscale}
\title{Inverse Standardisation}
\usage{
backscale(pars, y = NULL, coef = NULL)
}
\arguments{
\item{pars}{list as defined in section \emph{Value}, or \code{NULL} (if \code{family} is provided)}

\item{y}{\eqn{n_1}-dimensional response vector}

\item{coef}{\eqn{(1+p)-dimensional vector} containing the estimated intercept and the estimated slopes or \code{NULL} (default)}
}
\value{
Returns a list with slots \code{y_original} or \code{coef}.
}
\description{
Transforms response variable back to original scale or transforms coefficients for predictor variables and response variable on original scales.
}
\examples{

# simulate data
n <- 100; p <- 3
sd <- stats::rpois(n=p,lambda=5)
x <- sapply(X=sd,FUN=function(x) stats::rnorm(n=n,sd=x))
beta <- stats::rnorm(n=p)
y <- x \%*\% beta + stats::rnorm(n=n)

# regression without standardisation
lm1 <- stats::lm(y~x)
y_hat1 <- fitted(lm1)
coef1 <- coef(lm1)

# regression with standardisation
scale <- forescale(x=x,y=y,family="gaussian")
lm2 <- stats::lm(scale$y~scale$x)
result <- backscale(pars=scale$pars,y=fitted(lm2),coef=coef(lm2))
y_hat2 <- result$y_original
coef2 <- result$coef

# equality
all.equal(y_hat1,y_hat2)
all.equal(coef1,coef2,check.attributes=FALSE)

}
\seealso{
forescale
}
