% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{backscale}
\alias{backscale}
\title{Inverse Standardisation}
\usage{
backscale(pars, y = NULL, coef = NULL)
}
\arguments{
\item{pars}{list as defined in section \emph{Value},
or \code{NULL} (if \code{family} is provided)}

\item{y}{\eqn{n_1}-dimensional response vector}

\item{coef}{\eqn{(1 + p)-dimensional vector}
containing the estimated intercept
and the estimated slopes or \code{NULL} (default)}
}
\value{
Returns a list with slots \code{y_original} or \code{coef}.
}
\description{
Transforms response variable back to original scale
or transforms coefficients for predictor variables and response variable
on original scales.
}
\examples{
# simulate data
family <- "cox"
n0 <- 100; n1 <- 50; p <- 3
n <- n0 + n1
fold <- rep(c(0, 1), times = c(n0, n1))
sd <- stats::rpois(n = p, lambda = 5)
x <- data.frame(x = sapply(X = sd,
                          FUN = function(x) stats::rnorm(n = n, sd = x)))
beta <- stats::rnorm(n = p)
eta <- as.matrix(x) \%*\% beta
if (family == "gaussian") {
 y <- stats::rnorm(n = n, mean = eta)
} else if (family == "binomial") {
 y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (family == "poisson") {
 y <- stats::rpois(n = n, lambda = exp(eta))
} else if (family == "cox") {
 time <- stats::rexp(n = n, rate = exp(eta))
 status <- stats::rbinom(n = n, prob = 0.5, size = 1)
 y <- survival::Surv(time = time, event = status)
}

# regression without standardisation
if (family == "cox") {
 lm1 <- survival::coxph(y[fold == 0]~., data=x[fold == 0, ])
} else {
 lm1 <- stats::glm(y[fold == 0]~., data=x[fold == 0, ], family=family)
}
coef1 <- stats::coef(lm1)
yhat1 <- predict(lm1, newdata = x[fold == 1, ])

# regression with standardisation
scale <- forescale(x = x[fold == 0, ], y = y[fold == 0], family = family)
if (family == "cox") {
 lm2 <- survival::coxph(scale$y~., data = data.frame(scale$x))
} else {
 lm2 <- stats::glm(scale$y~., data = data.frame(scale$x), family = family)
}
coef_temp <- stats::coef(lm2)
newx_temp <- forescale(x = x[fold == 1, ], pars = scale$pars)$x
yhat_temp <- predict(object = lm2, newdata = data.frame(newx_temp))
result <- backscale(pars = scale$pars, y = yhat_temp, coef = coef_temp)
coef2 <- result$coef
yhat2 <- result$y_original

# equality
all.equal(coef1, coef2, check.attributes = FALSE)
all.equal(yhat1, yhat2)
}
\seealso{
\code{\link{forescale}()}
}
