% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_sign_prec}
\alias{calc_sign_prec}
\title{Calculates precision for sign variable}
\usage{
calc_sign_prec(truth, estim)
}
\arguments{
\item{truth}{integer vector with values in \eqn{\{-1, 0, 1\}}}

\item{estim}{integer vector of same length with values in \eqn{\{-1, 0, 1\}}}
}
\value{
Returns a scalar between 0 (minimum precision) and 1 (maximum precision),
or \code{NA} if all estimated signs equal 0.
}
\description{
Calculates precision for ternary variables with support \eqn{\{-1, 0, 1\}},
i.e., the proportion of positive or negative estimated signs
that match the true sign.
}
\examples{
truth <- sample(x = c(-1, 0, 1), size = 10, replace = TRUE)
estim <- sample(x = c(-1, 0, 1), size = 10, replace = TRUE)
calc_sign_prec(truth = truth, estim = estim) # observed value
calc_sign_prec(truth = truth, estim = -truth) # lower limit 0
calc_sign_prec(truth = truth, estim = truth) # upper limit 1
calc_sign_prec(truth = truth, estim = 0 * estim) # not defined

}
