% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.corila}
\alias{coef.cv.corila}
\title{Extract coefficients}
\usage{
\method{coef}{cv.corila}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{s}{character \code{"lambda.min"} or numeric non-negative scalar}

\item{...}{(for compatibility with \link[stats:coef]{stats::coef})}
}
\value{
Returns an \eqn{(1 + p)}-dimensional vector of the estimated coefficients.
The first entry is the estimated intercept,
and the other \eqn{p} entries are the estimated slopes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts coefficients from an object of class \code{"cv.corila"}.
}
\details{
This function calls \code{\link[=.combine_slopes]{.combine_slopes()}}
to combine positive and negative coefficients
and \code{\link[=.backscale]{.backscale()}} to bring coefficients back to the original scale.
}
\examples{
\donttest{
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = data$group,
                   primary = data$primary)
beta_hat <- coef(object = model)
y_hat <- predict(object = model, newx = data$x_test)
}

# example for automatic mutation testing (with the R package autotest)
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = as.double(data$group),
                   primary = data$primary,
                   alpha_init = 0.0,
                   threshold = 0.0,
                   foldid = rep(1:10, length.out = nrow(data$x_train)))

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link[=cv.corila]{cv.corila()}}
and make predictions with \link[=predict.cv.corila]{predict()}.
}
\keyword{methods}
