% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.corila}
\alias{coef.cv.corila}
\title{Extract coefficients}
\usage{
\method{coef}{cv.corila}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{s}{character \code{"lambda.min"} or numeric value}

\item{...}{(for compatibility with \link[stats:coef]{stats::coef})}
}
\value{
Returns an \eqn{(1 + p)}-dimensional vector of the estimated coefficients.
The first entry is the estimated intercept,
and the other \eqn{p} entries are the estimated slopes.
}
\description{
Extracts coefficients from an object of class \code{"cv.corila"}.
}
\details{
This function calls \code{\link[=.combine_slopes]{.combine_slopes()}}
to combine positive and negative coefficients
and \code{\link[=.backscale]{.backscale()}} to bring coefficients back to the original scale.
}
\examples{
#data <- simulate_data()

# option A: balanced and stratified folds (generated by corila:::.folds)
#foldid <- NULL

# option B: define folds manually
#foldid <- sample(seq_len(10), size = length(data$y_train), replace = TRUE)

#group <- seq_len(ncol(data$x_train))
#group <- LETTERS[data$group]

#model <- cv.corila(x = data$x_train,
#                   y = data$y_train,
#                   group = group,
#                   primary = data$primary,
#                   alpha_init = 0.0,
#                   tune = "none",
#                   foldid = foldid)

#beta_hat <- coef(object = model)
#y_hat <- predict(object = model, newx = data$x_test)

n <- 30
p <- 20
q <- 4
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(as.numeric(seq_len(q)), length.out = p)
primary <- rep(c(TRUE, FALSE), length.out = p)
nfolds <- 3
foldid <- rep(1:nfolds, length.out = n)
model <- cv.corila(x = x,
                    y = y,
                    alpha_init = 0.0,
                    group = group,
                    primary = primary,
                    tune = "none",
                    foldid = foldid)

#warning("Re-activate examples.")
#\dontshow{
#model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary,
#                    tune = "none")
# }
# \donttest{
# model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary)
# }

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link[=cv.corila]{cv.corila()}}
and make predictions with \link[=predict.cv.corila]{predict()}.
}
\keyword{methods}
