% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.cv.corila}
\alias{coef.cv.corila}
\title{Extract coefficients}
\usage{
\method{coef}{cv.corila}(object, s = "lambda.min")
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{s}{character \code{"lambda.min"} or numeric value}
}
\description{
Extracts coefficients from an object of class \code{"cv.corila"}.
}
\seealso{
Fit models with \code{\link{cv.corila}()} and make predictions with \code{\link{predict.cv.corila}()}.
}
