% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.corila}
\alias{coef.cv.corila}
\title{Extract coefficients}
\usage{
\method{coef}{cv.corila}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{s}{character \code{"lambda.min"} or numeric value}

\item{...}{(for compatibility with \link[stats:coef]{stats::coef})}
}
\value{
Returns an \eqn{(1 + p)}-dimensional vector of the estimated coefficients.
The first entry is the estimated intercept,
and the other \eqn{p} entries are the estimated slopes.
}
\description{
Extracts coefficients from an object of class \code{"cv.corila"}.
}
\details{
This function calls \code{\link[=.combine_slopes]{.combine_slopes()}}
to combine positive and negative coefficients
and \code{\link[=.backscale]{.backscale()}} to bring coefficients back to the original scale.
}
\examples{
data <- simulate_data()

# option A: balanced and stratified folds (generated by corila:::.folds)
foldid <- NULL

# option B: define folds manually
foldid <- sample(seq_len(10), size = length(data$y_train), replace = TRUE)

model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = data$group,
                   primary = data$primary,
                   alpha_init = 0.0,
                   tune = "none",
                   foldid = foldid)

warning("Re-activate examples.")
#\dontshow{
#model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary,
#                    tune = "none")
# }
# \donttest{
# model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary)
# }

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link[=cv.corila]{cv.corila()}}
and make predictions with \link[=predict.cv.corila]{predict()}.
}
\keyword{methods}
