% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.multiridge}
\alias{coef.multiridge}
\title{Extract Coefficients}
\usage{
\method{coef}{multiridge}(object, ...)
}
\arguments{
\item{object}{object of class \code{"multiridge"}}

\item{...}{(not used)}
}
\description{
Extracts coefficients from a multi-penalty ridge regression model.
}
\seealso{
\code{\link{multiridge}()}, \code{\link{predict.multiridge}()}
}
