% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ridge.R
\name{coef.multiridge}
\alias{coef.multiridge}
\title{Extract coefficients}
\usage{
\method{coef}{multiridge}(object, ...)
}
\arguments{
\item{object}{object of type \code{"multiridge"}}

\item{...}{(for compatibility with \link[stats:coef]{stats::coef})}
}
\value{
Returns an \eqn{(1 + p)}-dimensional vector of estimated coefficients
(estimated intercept and estimated slopes).
}
\description{
Extracts coefficients from a multi-penalty ridge regression model.
}
\examples{
# minimal example
n <- 50L; p <- 20L; q <- 5L
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
model <- multiridge(x = x, y = y, z = z)

# fitting with given folds
foldid <- sample(seq_len(10L), size = n, replace = TRUE)
model <- multiridge(x = x, y = y, z = z, foldid = foldid)

# fitting with given penalties
penalties <- abs(rnorm(q))
model <- multiridge(x = x, y = y, z = z, penalties = penalties)

\donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x \%*\% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
sapply(coef, function(x) mean((beta-x[-1])^2))

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
}

}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee}
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\doi{10.1080/10618600.2021.1904962}.
}
\seealso{
Fit models with \code{\link[=multiridge]{multiridge()}}
and make predictions with \link[=predict.multiridge]{predict()}.
}
\keyword{methods}
