% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{corila}
\alias{corila}
\alias{.construct_penalty_factors}
\title{Sparse group lasso regression (without cross-validation)}
\usage{
corila(
  x,
  y,
  group,
  primary,
  family,
  hyper,
  alpha_init,
  alpha_final,
  cor,
  foldid,
  nfolds,
  lambda_init,
  silent = FALSE
)

.construct_penalty_factors(coef, group, cor, names, primary, hyper)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{group}{group structure (multiple options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
}}

\item{primary}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE} for "primary predictors")
or must be excluded from the final model
(\code{FALSE} for "auxiliary predictors")}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{hyper}{list of \eqn{m}-dimensional vectors
or a data frame with \eqn{m} rows
containing candidate values
for the regularisation and mixing hyperparameters,
as well as for the correlation hard-thresholding hyperparameter}

\item{alpha_init}{A scalar specifying the method used for obtaining initial coefficients:
\itemize{
\item a numeric scalar in the unit interval
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
to define the mixing parameter for elastic net regression
(default: ridge penalisation with \code{alpha_init}=0);
\item the character scalar \code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"})
\item the character scalar\code{"multiridge"}
to use multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
\item \code{NA} to set all initial coefficients equal to 1
}}

\item{alpha_final}{elastic net mixing parameter for final regression:
numeric between 0 for ridge penalisation
and 1 for lasso penalisation
(default: lasso penalisation with \code{alpha_final}=1)}

\item{cor}{character string \code{"pearson"},
\code{"spearman"} (default),
or \code{"kendall"};
or a correlation matrix
(\eqn{p} rows, \eqn{p} columns,
entries between \eqn{-1} and \eqn{+1})}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers
(minimum \eqn{1}, maximum \code{nfolds})}

\item{nfolds}{positive integer specifying the number of folds
(minimum \eqn{3}, maximum \eqn{n})

NB: If \code{foldid} is provided, \code{nfolds} is overwritten by \code{max(foldid)}.}

\item{lambda_init}{regularisation hyperparameter(s),
or \code{NULL} (cross-validation)}

\item{silent}{Should messages from \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
be suppressed? (logical scalar, \code{FALSE} or \code{TRUE})}
}
\value{
Returns an object of class \code{"corila"}.
}
\description{
Fits an initial ridge regression to obtain weights
for an adaptive lasso regression
that allows for heterogeneous, overlapping and unknown groups
of correlated variables.
}
\details{
The numbers of observations (samples) for training or testing
are indicated by \eqn{n_0} and \eqn{n_1}, respectively,
the number of predictors (features) is indicated by \eqn{p},
and the number of predictor group is indicated by \eqn{q}.
Observations are indexed by \eqn{i} in \eqn{\{1, \ldots, n\}},
predictors are indexed by \eqn{j} in \eqn{\{1, \ldots, p\}},
and predictor groups are indexed by \eqn{k} in \eqn{\{1, \ldots, q\}}.
The number of predictors in the \eqn{k^{\text{th}}} group
is indicated by \eqn{p_k},
with \eqn{\sum_{k=1}^q p_k = p} for non-overlapping groups.
}
\examples{
\donttest{
\dontshow{corila <- corila:::corila}
# simulation
n <- 100L
p <- 50L
group <- rep(x = seq_len(10L), each = 5L)
primary <- rep(x = TRUE, times = p)
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n = n)

# model fitting
hyper <- data.frame(wgt_local = 0.5, exp_local = 1.0,
                    wgt_global = 0.5, exp_global = 1.0,
                    threshold = 0.0)
object <- corila(x = x,
                 y = y,
                 group = group,
                 primary = primary,
                 family = "gaussian",
                 alpha_init = 0.0,
                 alpha_final = 1.0,
                 cor = "spearman",
                 foldid = NULL,
                 nfolds = 10L,
                 hyper = hyper,
                 lambda_init = NULL)

y_hat <- stats::predict(object, newx = x, index = 1L, s = 0.0)
}

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Estimate parameters and tune hyperparameters (using cross-validation)
with \code{\link[=cv.corila]{cv.corila()}}.
Make predictions for a range of hyperparameters
with \link[=predict.corila]{predict()}.

This function calls
\code{\link[=.forescale]{.forescale()}} and \code{\link[=.backscale]{.backscale()}}
for standardising data and bringing results back to the original scale,
respectively,
\code{\link[=.folds]{.folds()}} for splitting samples into folds,
\code{\link[=.estim_initial_coefs]{.estim_initial_coefs()}} for obtaining initial coefficients,
\code{\link[=.is_adjacent]{.is_adjacent()}} for identifying adjacent predictors,
and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} and \code{\link[glmnet:glmnet]{glmnet::glmnet()}}
for adaptive lasso regression.
}
\keyword{internal}
