% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corila}
\alias{corila}
\title{Group lasso}
\usage{
corila(
  x,
  y,
  group,
  type,
  family,
  hyper,
  cor = "spearman",
  cond = NULL,
  lambda.com = NULL,
  lambda.sep = NULL,
  lambda.ind = NULL,
  trial = TRUE,
  fuse = "mean",
  init.multi = FALSE
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices (in \eqn{\{1,\ldots,q\}}) or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices (in \eqn{\{1,\ldots,p\}}) or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{th}}} row and the \eqn{k^{\text{th}}} column indicates whether information should be transferred from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{hyper}{list of of \eqn{m}-dimensional vectors or a data frame with \eqn{m} rows containing candidate values for hyperparameters}

\item{cor}{character string \code{"pearson"}, \code{"spearman"} (default), or \code{"kendall"}; or \eqn{p \times p} correlation matrix}

\item{cond}{\code{NULL}}

\item{lambda.com, lambda.sep, lambda.ind}{regularisation parameters, or \code{NULL} (cross-validation)}

\item{trial}{logical}

\item{fuse}{character string \code{"mean"} for arithmetic mean  or \code{"pca"} for first principal component}

\item{init.multi}{Use multi-penalty ridge regression (one penalty for each group) to obtain initial coefficients? (\code{TRUE} or \code{FALSE})}
}
\value{
Returns an object of class \code{"corila"}.
}
\description{
Fits an initial ridge regression to obtain weights for an adaptive lasso 
regression that allows for heterogeneous, overlapping and unknown groups of correlated variables.
}
\examples{
\donttest{
# simulation
n <- 100
p <- 50
group <- rep(x=1:10,each=5)
type <- rep(x=1,times=length(group))
x <- matrix(data=rnorm(n*p),nrow=n,ncol=p)
y <- rnorm(n=n)

# model fitting
hyper <- data.frame(exp.local=1,wgt.local=0.5,exp.global=1,wgt.global=0.5)
object <- corila(x,y,group,type,family="gaussian",hyper=hyper)

y_hat <- predict(object,newx=x,index=1,s=0)
}
}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Estimate parameters and tune hyperparameters (using cross-validation) with \code{\link{cv.corila}()}. Make predictions for a range of hyperparameters with \code{\link{predict.corila}()}.

This function calls \code{\link{forescale}} and \code{\link{backscale}} for standardising data and bringing results back to the original scale, respectively, \code{\link{multiridge}} for obtaining initial group penalties, and \code{\link[glmnet]{cv.glmnet}} and \code{\link[glmnet]{glmnet}} for adaptive lasso regression.
}
