% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corila}
\alias{corila}
\title{Group lasso}
\usage{
corila(
  x,
  y,
  group,
  type,
  family,
  hyper,
  cor = "spearman",
  cond = NULL,
  lambda.com = NULL,
  lambda.sep = NULL,
  lambda.ind = NULL,
  trial = TRUE,
  mode = "mean",
  init.multi = FALSE
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{group}{XXX}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string "gaussian", "binomial", "poisson" or "cox"}

\item{hyper}{list of of \eqn{m}-dimensional vectors or a data frame with $m$ rows containing candidate values for hyperparameters}

\item{cor}{character string \code{"pearson"}, \code{"spearman"} (default), or \code{"kendall"}; or \eqn{p \times p} correlation matrix}

\item{cond}{\code{NULL}}

\item{lambda.com, lambda.sep, lambda.ind}{\code{NULL}}

\item{trial}{logical}

\item{mode}{character string "mean" for arithmetic mean  or "pca" for first principal component}

\item{init.multi}{logical}
}
\value{
Returns an object of class \code{"corila"}.
}
\description{
Fits an initial ridge regression to obtain weights for an adaptive lasso 
regression that allows for heterogeneous, overlapping and unknown groups of correlated variables.
}
\examples{
NULL

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Estimate parameters and tune hyperparameters (using cross-validation) with \code{\link{cv.corila}()}. Make predictions for a range of hyperparameters with \code{\link{predict.corila}()}.
}
