% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crossval}
\alias{crossval}
\title{Cross-validation method}
\usage{
crossval(
  x,
  y,
  family,
  group = NULL,
  type = NULL,
  iter = 5,
  nfolds = 10,
  init.multi = FALSE,
  trial = FALSE,
  method = NULL
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"} or \code{"cox"}}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{(th)}}} row and the \eqn{k^{\text{(th)}}} column indicates whether information should be transferred from the \eqn{j^{\text{(th)}}} to the \eqn{k^{\text{(th)}}} variable}

\item{type}{\eqn{p}-dimensional vector}

\item{iter}{number of cross-validation iterations}

\item{nfolds}{number of cross-validation folds}

\item{init.multi}{logical}

\item{trial}{logical}

\item{method}{character vector listing all methods to be compared}
}
\value{
Returns a list with the following slots:
\itemize{
\item \code{nzero} non-zero coefficients
\item \code{metric} metric
}
Both slot contain a data frame with one row for each iteration (\code{iter}) and one column for each \code{method}.
}
\description{
Compares methods with cross-validation method
}
\details{
This function implements repeated \eqn{k}-fold cross-validation (e.g., 5 repetitions of 10-fold cross-validation).
}
\examples{
n <- 100
p <- 20
x <- matrix(rnorm(n*p),nrow=n,ncol=p)
y <- stats::rnorm(n)
results <- crossval(x,y,family="gaussian",method=c("mean","corila"),trial=TRUE)

}
