% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{cv.corila}
\alias{cv.corila}
\title{Sparse group lasso regression}
\usage{
cv.corila(
  x,
  y,
  group,
  primary = NULL,
  family = "gaussian",
  alpha_init = 0,
  cor = "spearman",
  alpha_final = 1,
  nfolds = 10L,
  foldid = NULL,
  tune = "weight",
  na_action = "error",
  silent = FALSE
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{group}{group structure (multiple options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
(usually with a unit diagonal so that predictors retain the information
from their initial coefficient)
}}

\item{primary}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE} for "primary predictors")
or must be excluded from the final model
(\code{FALSE} for "auxiliary predictors")}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{alpha_init}{A scalar specifying the method used for obtaining initial coefficients:
\itemize{
\item a numeric scalar in the unit interval
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
to define the mixing parameter for elastic net regression
(default: ridge penalisation with \code{alpha_init}=0);
\item the character scalar \code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"})
\item the character scalar\code{"multiridge"}
to use multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
\item \code{NA} to set all initial coefficients equal to 1
}}

\item{cor}{character string \code{"pearson"},
\code{"spearman"} (default),
or \code{"kendall"};
or a correlation matrix
(\eqn{p} rows, \eqn{p} columns,
entries between \eqn{-1} and \eqn{+1})}

\item{alpha_final}{elastic net mixing parameter for final regression:
numeric between 0 for ridge penalisation
and 1 for lasso penalisation
(default: lasso penalisation with \code{alpha_final}=1)}

\item{nfolds}{positive integer specifying the number of folds
(minimum \eqn{3}, maximum \eqn{n})

NB: If \code{foldid} is provided, \code{nfolds} is overwritten by \code{max(foldid)}.}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers
(minimum \eqn{1}, maximum \code{nfolds})}

\item{tune}{character string for determining the candidate values
for the hyperparameters:
\itemize{
\item \code{"none"}:
fixed weights and exponents
(\code{wgt_local}=1, \code{exp_local}=1, \code{wgt_global}=0),
no tuning
(share information within groups, ignore correlations)
\item \code{"weight"} (default):
fixed exponents (\code{exp_local}=0, \code{exp_global}=1),
tuning \code{wgt_local}=1-\code{wgt_global}
(find a compromise between sharing information within groups
and between correlated predictors)
\item \code{"exponent"}:
fixed weights (\code{wgt_local}=1, \code{wgt_global}=0),
tuning \code{exp_local}
(share information between correlated predictors in the same group,
determine level of trust in correlation coefficients)
\item \code{"bivariate"}:
tuning \code{wgt_local}=1-\code{wgt_global} and \code{exp_local}=\code{exp_global}
(find a compromise between sharing information
between predictors in the same group and between all predictors,
determine level of trust in correlation coefficients)
\item \code{"factorial"}:
tuning \code{wgt_local}, \code{exp_local}, \code{wgt_global}, \code{exp_global}
(unrestricted information sharing
with weights possibly not summing to one
and possibly different exponents)
}

(The internal function \code{\link[=.set_candidates]{.set_candidates()}} uses this argument
to create a grid of candidates values for
\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global}.)}

\item{na_action}{character \code{"error"} to trigger an error
if any observation has a missing predictor or a missing response
or \code{"complete_cases"} to exclude observations
with a missing predictor or a missing response from model fitting
(while providing fitted values for these observations)}

\item{silent}{Should messages from \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
be suppressed? (logical scalar, \code{FALSE} or \code{TRUE})}
}
\value{
Returns an object of class \code{"cv.corila"},
a list with the following slots:
\itemize{
\item \code{model}:
list with one slot for each combination of hyperparameters,
each slot contains an object of class \code{"glmnet"}
\item \code{hyper}:
data frame with one row for each combination of hyperparameters,
four columns for the values of the hyperparameters
(\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global})
and a column for the cross-validated loss (\code{cvm})
\item \code{id_hyper}:
index of combination of hyperparameters
leading to the lowest cross-validated loss
\item \code{lambda.min}
optimised regularisation hyperparameter
\item \code{scale}:
output from \code{\link[=.forescale]{.forescale()}}
\item \code{y_hat}:
\eqn{n_0}-dimensional vector of fitted values
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Optimises the parameters and the hyperparameters of the sparse group lasso.
}
\details{
The numbers of observations (samples) for training or testing
are indicated by \eqn{n_0} and \eqn{n_1}, respectively,
the number of predictors (features) is indicated by \eqn{p},
and the number of predictor group is indicated by \eqn{q}.
Observations are indexed by \eqn{i} in \eqn{\{1, \ldots, n\}},
predictors are indexed by \eqn{j} in \eqn{\{1, \ldots, p\}},
and predictor groups are indexed by \eqn{k} in \eqn{\{1, \ldots, q\}}.
The number of predictors in the \eqn{k^{\text{th}}} group
is indicated by \eqn{p_k},
with \eqn{\sum_{k=1}^q p_k = p} for non-overlapping groups.
}
\examples{
\donttest{
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = data$group,
                   primary = data$primary)
beta_hat <- coef(object = model)
y_hat <- predict(object = model, newx = data$x_test)
}

# example for automatic mutation testing (with the R package autotest)
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = as.double(data$group),
                   primary = data$primary,
                   alpha_init = 0.0,
                   foldid = rep(x = seq_len(10L),
                                length.out = nrow(data$x_train)))

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Extract coefficients with \link[=coef.cv.corila]{coef()}
and make predictions with \link[=predict.cv.corila]{predict()}.

This user function repeatedly calls \code{\link[=corila]{corila()}}
with different values for the regularisation and mixing hyperparameters.

The arguments of this function are validated
with the helper functions listed \link[=validate]{here}.
}
\keyword{classif}
\keyword{methods}
\keyword{models}
\keyword{regression}
