% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.corila}
\alias{cv.corila}
\title{Sparse Group Lasso}
\usage{
cv.corila(
  x,
  y,
  group,
  type = NULL,
  family = "gaussian",
  cor = "spearman",
  fuse = "mean",
  init.multi = FALSE,
  trial = TRUE,
  foldid = NULL
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{(th)}}} row and the \eqn{k^{\text{(th)}}} column indicates whether information should be transferred from the \eqn{j^{\text{(th)}}} to the \eqn{k^{\text{(th)}}} variable}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"} or \code{"cox"}}

\item{cor}{character string \code{"pearson"}, \code{"spearman"} (default), or \code{"kendall"}; or \eqn{p \times p} correlation matrix}

\item{fuse}{character string \code{"mean"} for arithmetic mean  or \code{"pca"} for first principal component}

\item{init.multi}{logical}

\item{trial}{logical}

\item{foldid}{\eqn{n}-dimensional vector containing the fold identifiers}
}
\description{
Optimises the parameters and the hyperparameters of the sparse group lasso.
}
\examples{
1+1

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Extract coefficients with \code{\link[=coef.cv.corila]{coef}()} and make predictions with \code{\link[=predict.cv.corila]{predict}()}.
}
