% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.corila}
\alias{cv.corila}
\title{Sparse Group Lasso}
\usage{
cv.corila(
  x,
  y,
  group,
  type = NULL,
  family = "gaussian",
  cor = "spearman",
  mode = "mean",
  init.multi = FALSE,
  trial = TRUE,
  foldid = NULL
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{group}{XXX}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string "gaussian", "binomial", "poisson" or "cox"}

\item{cor}{character string \code{"pearson"}, \code{"spearman"} (default), or \code{"kendall"}; or \eqn{p \times p} correlation matrix}

\item{mode}{character string "mean" for arithmetic mean  or "pca" for first principal component}

\item{init.multi}{logical}

\item{trial}{logical}

\item{foldid}{\eqn{n}-dimensional vector containing the fold identifiers}
}
\description{
Optimises the parameters and the hyperparameters of the sparse group lasso.
}
\seealso{
[corila()]
}
