% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.corila}
\alias{cv.corila}
\title{Sparse Group Lasso}
\usage{
cv.corila(
  x,
  y,
  group,
  type = NULL,
  family = "gaussian",
  cor = "spearman",
  fuse = "mean",
  init.multi = FALSE,
  trial = TRUE,
  tune = "both",
  foldid = NULL
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n}-dimensional response vector}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices (in \eqn{\{1,\ldots,q\}}) or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices (in \eqn{\{1,\ldots,p\}}) or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{th}}} row and the \eqn{k^{\text{th}}} column indicates whether information should be transferred from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{cor}{character string \code{"pearson"}, \code{"spearman"} (default), or \code{"kendall"}; or \eqn{p \times p} correlation matrix}

\item{fuse}{character string \code{"mean"} for arithmetic mean  or \code{"pca"} for first principal component}

\item{init.multi}{Use multi-penalty ridge regression (one penalty for each group) to obtain initial coefficients? (\code{TRUE} or \code{FALSE})}

\item{trial}{logical}

\item{tune}{character \code{"wgt"}, \code{"exp"}, or \code{"both"} for determining the candidate values for the hyperparameters; or list with slots \code{wgt.local}, \code{wgt.global}, \code{exp.local}, and \code{exp.global} (not yet implemented)}

\item{foldid}{\eqn{n}-dimensional vector containing the fold identifiers}
}
\value{
Returns an object of class \code{cv.corila}.
}
\description{
Optimises the parameters and the hyperparameters of the sparse group lasso.
}
\examples{
\donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100,50)
z <- rep(x=seq_along(p),times=p)
x <- sapply(X=z,FUN=function(x) stats::rnorm(n=n,sd=x))
beta <- stats::rnorm(n=sum(p),mean=1,sd=0)*stats::rbinom(n=sum(p),size=1,prob=0.2)
eta <- x \%*\% beta
family <- "gaussian"
if(family=="gaussian"){
 y <- eta + 0.5*stats::rnorm(n=n,sd=stats::sd(eta))
} else if(family=="binomial"){
 y <- stats::rbinom(n=n,size=1,prob=1/(1+exp(-eta)))
} else if(family=="cox"){
 time <- stats::rexp(n=n,rate=exp(eta))
 status <- stats::rbinom(n=n,prob=0.5,size=1)
 y <- survival::Surv(time=time,event=status)
}
cond <- rep(x=c(TRUE,FALSE),times=c(n0,n1))

y_hat <- coef <- list()

# standard lasso regression
object <- glmnet::cv.glmnet(x=x[cond,],y=y[cond],family=family,alpha=1)
coef$glmnet <- stats::coef(object=object,s="lambda.min")
y_hat$glmnet <- stats::predict(object=object,newx=x[!cond,],type="response",s="lambda.min")

# flexible group lasso regression
object <- cv.corila(x=x[cond,],y=y[cond],group=z,family=family)
coef$corila <- stats::coef(object=object)
y_hat$corila <- stats::predict(object=object,newx=x[!cond,])

# estimation performance
sapply(coef,function(x) stats::cor(beta,x[-1]))
sapply(coef,function(x) mean((beta-x[-1])^2))

# predictive performance
if(family=="gaussian"){
 metric <- sapply(X=y_hat,FUN=function(x) mean((x-y[!cond])^2))
} else if(family=="binomial"){
 metric <- sapply(X=y_hat,FUN=function(x) pROC::auc(response=y[!cond],predictor=as.vector(x),levels=c(0,1),direction="<"))
} else if(family=="cox"){
 metric <- sapply(X=y_hat,FUN=function(x) survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
}
}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Extract coefficients with \code{\link[=coef.cv.corila]{coef}()} and make predictions with \code{\link[=predict.cv.corila]{predict}()}.

This user function repeatedly calls \code{\link{corila}} with different values for the hyperparameters.
}
