% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{cv.corila}
\alias{cv.corila}
\title{Sparse group lasso regression}
\usage{
cv.corila(
  x,
  y,
  group,
  primary = NULL,
  alpha_init = 0,
  alpha_final = 1,
  family = "gaussian",
  nfolds = 10L,
  cor = "spearman",
  tune = "weight",
  foldid = NULL,
  na_action = "error",
  silent = FALSE
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{group}{group structure (three options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
}}

\item{primary}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE} for "primary predictors")
or must be excluded from the final model
(\code{FALSE} for "auxiliary predictors")}

\item{alpha_init}{elastic net mixing parameter
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
for initial regression
(default: ridge penalisation with \code{alpha_init}=0);
alternative choices are
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"}),
\code{"multiridge"} for multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
or \code{NA} to set all initial coefficients equal to 1}

\item{alpha_final}{elastic net mixing parameter for final regression:
numeric between 0 for ridge penalisation
and 1 for lasso penalisation
(default: lasso penalisation with \code{alpha_final}=1)}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{nfolds}{positive integer specifying the number of folds
(minimum \eqn{3}, maximum \eqn{n})}

\item{cor}{character string \code{"pearson"},
\code{"spearman"} (default),
or \code{"kendall"};
or \eqn{p \times p} correlation matrix}

\item{tune}{character string for determining the candidate values
for the hyperparameters:
\itemize{
\item \code{"none"}:
fixed weights and exponents
(\code{wgt_local}=1, \code{exp_local}=1, \code{wgt_global}=0),
no tuning
\item \code{"weight"}:
fixed exponents (\code{exp_local}=0, \code{exp_global}=1),
tuning \code{wgt_local}=1-\code{wgt_global}
\item \code{"exponent"}:
fixed weights (\code{wgt_local}=1, \code{wgt_global}=0),
tuning \code{exp_local}
\item \code{"bivariate"}:
tuning \code{wgt_local}=1-\code{wgt_global} and \code{exp_local}=\code{exp_global}
\item \code{"factorial"}:
tuning \code{wgt_local}, \code{exp_local}, \code{wgt_global}, \code{exp_global}
}

(to implement: list with slots
\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global})}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers}

\item{na_action}{character \code{"error"} to trigger an error
if any observation has a missing predictor or a missing response
or \code{"complete_cases"} to exclude observations
with a missing predictor or a missing response from model fitting
(while providing fitted values for these observations)}

\item{silent}{Should messages from \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
be suppressed? logical}
}
\value{
Returns an object of class \code{"cv.corila"},
a list with the following slots:
\itemize{
\item \code{model}:
list with one slot for each combination of hyperparameters,
each slot contains an object of class \code{"glmnet"}
\item \code{hyper}:
data frame with one row for each combination of hyperparameters,
four columns for the values of the hyperparameters
(\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global})
and a column for the cross-validated loss (\code{cvm})
\item \code{id_hyper}:
index of combination of hyperparameters
leading to the lowest cross-validated loss
\item \code{lambda.min}
optimised regularisation hyperparameter
\item \code{scale}:
output from \code{\link[=.forescale]{.forescale()}}
\item \code{y_hat}:
\eqn{n}-dimensional vector of fitted values
}
}
\description{
Optimises the parameters and the hyperparameters of the sparse group lasso.
}
\details{
The number of observations (samples) for training or testing
are indicated by \eqn{n_0} and \eqn{n_1}, respectively,
the number of variables (features) is indicated by \eqn{p},
and the number of variable groups is indicated by \eqn{q}.
Observations (samples) are indexed by \eqn{i} in \eqn{\{1, \ldots, n\}},
variables (features) are indexed by \eqn{j} in \eqn{\{1, \ldots, p\}},
and variable groups are indexed by \eqn{k} in \eqn{\{1, \ldots, q\}}.
The number of variables in the \eqn{k^{\text{th}}} group
is indicated by \eqn{p_k},
with \eqn{\sum_{k=1}^q p_k = p} for non-overlapping groups.
}
\examples{
# minimal example
set.seed(1)
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
cv.corila(x = x, y = y, group = group, primary = primary, tune = "none")

\donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x \%*\% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard lasso regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                            family = family, alpha = 1)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                               type = "response", s = "lambda.min")

# flexible group lasso regression
object <- cv.corila(x = x[cond, ], y = y[cond], group = z, family = family)
coef$corila <- stats::coef(object = object)
y_hat$corila <- stats::predict(object = object, newx = x[!cond, ])

# selection performance
sapply(coef, function(x) mean(sign(x[-1]) == sign(beta)))
sapply(coef, function(x) {
  sum(sign(x[-1]) != 0 & sign(x[-1]) == sign(beta)) / sum(x[-1] != 0)
})

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric

# privileged information
#primary <- stats::rbinom(n = sum(p), size = 1, prob = 0.5) == 1
#object <- cv.corila(x = x[cond, ], y = y[cond], group = z,
#                     primary = primary, family = family)
}

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Extract coefficients with \link[=coef.cv.corila]{coef()}
and make predictions with \link[=predict.cv.corila]{predict()}.

This user function repeatedly calls \code{\link[=corila]{corila()}}
with different values for the regularisation and mixing hyperparameters.
}
\keyword{classif}
\keyword{methods}
\keyword{models}
\keyword{regression}
