% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example data}
\format{
\subsection{\code{data}}{

A list with multiple slots:
\itemize{
\item \code{x_train}:
predictor matrix of the training observations
(\eqn{n_0} rows, \eqn{p} columns)
\item \code{y_train}:
response vector of the training observations
(length \eqn{n_0})
\item \code{group}:
integer vector indicating the group of the predictors
(length \eqn{p})
\item \code{primary}:
logical vector indicating
primary (\code{TRUE}) and auxiliary (\code{FALSE}) predictors
(length \eqn{p})
\item \code{beta}:
numeric vector of the effects of the predictors on the response
(length \eqn{p})
\item \code{x_test}:
\eqn{n_1 \times p} predictor matrix for the test observations
\item \code{y_test}:
response vector for the test observations of length \eqn{n_1}
}
}
}
\usage{
data
}
\description{
This is example data for modelling a response based on
grouped and correlated primary and auxiliary predictors.
}
\details{
Use the objects \code{x_train}, \code{y_train}, \code{group}, and \code{primary}
for model training.
Estimated coefficients can be compared with \code{beta}.
Use the object \code{x_test} for model testing.
Predicted values can be compared with \code{y_test}.
}
\keyword{datasets}
