% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{deviance.cv.corila}
\alias{deviance.cv.corila}
\title{Deviance}
\usage{
\method{deviance}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns a scalar.
}
\description{
Calculates the deviance.
}
\details{
Returns the deviance calculated by \code{\link[glmnet:deviance.glmnet]{glmnet::deviance.glmnet()}}
for the model with the optimised mixing and regularisation hyperparameters.
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# using S3 methods
n <- 10; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)

coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
\seealso{
The internal function \code{\link[=.deviance]{.deviance()}} calculates
the deviance from fitted and observed values.
}
