% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.assert}
\alias{.assert}
\title{Assertions}
\usage{
.assert(
  x = NULL,
  type = "numeric",
  dim = 1L,
  na.rm = FALSE,
  support = NULL,
  family = NULL,
  min = -Inf,
  max = Inf
)
}
\arguments{
\item{x}{scalar, vector, matrix, or array to be checked}

\item{type}{character \code{"numeric"} (default), \code{"integer"},
\code{"nominal"}, or \code{"logical"}}

\item{dim}{vector containing positive integers defining the dimensionality:
\code{dim = 1L} for a scalar,
\code{dim = Inf} for a vector of arbitrary length,
\code{dim = c(Inf, Inf)} for a matrix of arbitrary dimensions,
\code{dim = c(Inf, Inf, Inf)} for an array of arbitrary dimensions,
\code{dim = 100L} for a vector of length 100,
\code{dim = c(Inf, 100L)} for a matrix with 100 columns, etc.}

\item{na.rm}{logical;
\code{FALSE}: missing values are not allowed,
\code{TRUE}: missing values are allowed}

\item{support}{character vector (only used for \code{type = "nominal"})}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{min}{numerical value (not used for \code{type = "nominal"})}

\item{max}{numerical value (not used for \code{type = "nominal"})}
}
\value{
Returns \code{NULL} invisibly, or an error message.
}
\description{
Check whether provided arguments satisfy expectations.
}
\details{
This function is called by multiple function of the \link{corila-package}.
}
\examples{
\dontshow{.assert <- corila:::.assert}
.assert(x = NULL)
.assert(x = rnorm(n = 1L))
.assert(x = "A", type = "nominal", support = LETTERS)
.assert(x = rexp(n= 10L), dim = Inf, type = "numeric", min = 0.0)
.assert(x = c(NA, rpois(n = 9L, lambda = 4.0)), dim = 10L,
       type = "integer", na.rm = TRUE)
.assert(x = NA, na.rm = TRUE)
.assert(x = 1.0, na.rm = FALSE)
.assert(x = rpois(n = 10L, lambda = 4.0), dim = Inf,
       family = "poisson")

}
\seealso{
The function \code{\link[=.validate]{.validate()}} verifies whether the main arguments
have compatible dimensions (number of samples and features).
}
\keyword{internal}
