% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.assert}
\alias{.assert}
\title{Assertions}
\usage{
.assert(
  x,
  type,
  dim = 1,
  na.rm = FALSE,
  support = NULL,
  family = NULL,
  min = -Inf,
  max = Inf
)
}
\arguments{
\item{x}{scalar, vector, matrix, or array to be checked}

\item{type}{character \code{"numeric"}, \code{"integer"},
\code{"nominal"}, or \code{"logical"}}

\item{dim}{dimensionality:
\code{dim = 1} for a scalar,
\code{dim = Inf} for a vector of arbitrary length,
\code{dim = c(Inf, Inf)} for a matrix of arbitrary dimensions,
\code{dim = c(Inf, Inf, Inf)} for an array of arbitrary dimensions,
\code{dim = 100} for a vector of length 100,
\code{dim = c(Inf, 100)} for a matrix with 100 columns, etc.}

\item{na.rm}{logical;
\code{FALSE}: missing values are not allowed,
\code{TRUE}: missing values are allowed}

\item{support}{character vector (only used for \code{type = "nominal"})}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{min}{numerical value (not used for \code{type = "nominal"})}

\item{max}{numerical value (not used for \code{type = "nominal"})}
}
\value{
Returns \code{NULL} invisibly, or an error message.
}
\description{
Check whether provided arguments satisfy expectations.
}
\details{
This function is called by multiple function of the \link{corila-package}.
}
\examples{
corila:::.assert(x = NULL)
corila:::.assert(x = rnorm(1), type = "numeric")
corila:::.assert(x = "A", type = "nominal", support = LETTERS)

}
\seealso{
The function \code{\link[=.validate]{.validate()}} verifies whether the main arguments
have compatible dimensions (number of samples and features).
}
\keyword{internal}
