% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.combine_slopes}
\alias{.combine_slopes}
\title{Combine coefficients}
\usage{
.combine_slopes(alpha, beta)
}
\arguments{
\item{alpha}{estimated intercept:
scalar}

\item{beta}{estimated slopes:
numeric vector of length \eqn{2 * p} with non-negative entries,
namely of \eqn{p} estimated coefficients for positive effects
and \eqn{p} estimated coefficients for negative effects}
}
\value{
numeric vector of length \eqn{1 + p}
}
\description{
Combine estimated coefficients for positive effects
and estimated coefficients for negative effects.
}
\examples{
p <- 10
alpha <- rnorm(1)
temp <- rnorm(p)
beta <- pmax(c(temp, -temp), 0)
corila:::.combine_slopes(alpha = alpha, beta = beta)

}
\keyword{internal}
