% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.deviance}
\alias{.deviance}
\title{Deviance}
\usage{
.deviance(y, y_hat, family)
}
\arguments{
\item{y}{response:
numeric vector of length \code{n}}

\item{y_hat}{predicted response:
numeric vector of length \code{n}}

\item{family}{character}
}
\description{
Calculate the deviance
}
\examples{
# move these examples to unit tests (adding cox):
y <- rnorm(10)
.deviance(y = y , y_hat = y, family = "gaussian")

y <- rbinom(n = 10, size = 1, prob = 0.5)
.deviance(y = y , y_hat = y, family = "binomial")

y <- rpois(n = 10, lambda = 4)
.deviance(y = y , y_hat = y, family = "poisson")

}
\keyword{internal}
