% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.estim_initial_coefs}
\alias{.estim_initial_coefs}
\title{Initial coefficients}
\usage{
.estim_initial_coefs(x, y, family, alpha, group, foldid, nfolds, lambda)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector,
where \eqn{n_0} is the number of observations used for model training}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\emph{(ii)} list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
or \emph{(iii)} \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{foldid}{\eqn{n}-dimensional vector containing the fold identifiers}

\item{nfolds}{integer specifying the number of folds}
}
\description{
Estimate initial coefficients.
}
\examples{
n <- 20
p <- 10
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
beta <- rbinom(n = p, size = 1, prob = 0.5) * rnorm(p)
y <- drop(x \%*\% beta)
.estim_initial_coefs(x = x,
                     y = y,
                     family = "gaussian",
                     alpha = "spearman",
                     group = NULL,
                     foldid = NULL,
                     nfolds = 10,
                     lambda = NULL)

}
\keyword{internal}
