% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.estim_initial_coefs}
\alias{.estim_initial_coefs}
\title{Initial coefficients}
\usage{
.estim_initial_coefs(x, y, family, alpha_init, group, foldid, nfolds, lambda)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector,
where \eqn{n_0} is the number of observations used for model training}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{alpha_init}{elastic net mixing parameter
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
for initial regression
(default: ridge penalisation with \code{alpha_init}=0);
alternative choices are
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"}),
\code{"multiridge"} for multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
or \code{NA} to set all initial coefficients equal to 1}

\item{group}{group structure (three options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
}}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers}

\item{nfolds}{integer specifying the number of folds}

\item{lambda}{numeric scalar, or \code{NULL}
(determined by cross-validation)}
}
\description{
Estimate initial coefficients.
}
\details{
This function is called by \code{\link[=corila]{corila()}}.
It calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} or \code{\link[glmnet:glmnet]{glmnet::glmnet()}}
for an initial lasso, ridge, or elastic net regression,
\code{\link[=multiridge]{multiridge()}} for an initial multi-penalty ridge regression,
or \code{\link[stats:cor]{stats::cor()}} for initial correlation coefficients.
}
\examples{
n <- 20
p <- 10
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
beta <- rbinom(n = p, size = 1, prob = 0.5) * rnorm(p)
y <- drop(x \%*\% beta)
corila:::.estim_initial_coefs(x = x,
                              y = y,
                              family = "gaussian",
                              alpha_init = "spearman",
                              group = NULL,
                              foldid = NULL,
                              nfolds = 10,
                              lambda = NULL)

}
\keyword{internal}
