% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.expand_auxiliary}
\alias{.expand_auxiliary}
\title{Expand auxiliary features}
\usage{
.expand_auxiliary(x, primary)
}
\arguments{
\item{x}{matrix with \eqn{n} rows and either
\eqn{p_0} or \eqn{p_0 + p_1} features}

\item{primary}{logical vector of length \eqn{p_0 + p_1}
with \eqn{p_0} entries equal to \code{TRUE} (primary features)
and \eqn{p_1} entries equal to \code{FALSE} (auxiliary features)}
}
\value{
Returns a matrix with \eqn{n} rows and \eqn{p_0 + p_1} columns.
}
\description{
Add empty columns for auxiliary features.
}
\examples{
n <- 5
p <- 10
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
x_primary <- x[, primary]
x_expanded <- corila:::.expand_auxiliary(x = x_primary, primary = primary)
all(x_expanded[, primary] == x[, primary])
all(x_expanded[, !primary] == 0)

}
\seealso{
This function is called by \link[=predict.cv.corila]{predict()}.
}
\keyword{internal}
