% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.is_adjacent}
\alias{.is_adjacent}
\title{Adjacency}
\usage{
.is_adjacent(group, j, p, names)
}
\arguments{
\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\emph{(ii)} list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
or \emph{(iii)} \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{j}{index of predictor}

\item{p}{number of predictors}

\item{names}{names of predictors}
}
\value{
Returns a logical vector of length \eqn{p}.
}
\description{
Identifies adjacent predictors.
}
\examples{
p <- 5
names <- paste0("x", seq_len(p))
group <- list()
group$index_vector <- setNames(object = c(1, 1, 2, 2, 3), nm = names)
group$label_vector <- setNames(object = LETTERS[group$index_vector],
                                 nm = names(group$index_vector))
group$index_list <- lapply(X = setNames(nm = unique(group$label_vector)),
                     FUN = function(x) which(group$label_vector == x))
group$label_list <- lapply(group$index_list, names)
group$matrix <- 1 * outer(X = group$index_vector,
                          Y = group$index_vector,
                          FUN = "==")
corila:::.is_adjacent(group = group[[1]], j = 3, p = p, names = names)

}
\keyword{internal}
