% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mean_function}
\alias{.mean_function}
\title{Mean function}
\usage{
.mean_function(x, family)
}
\arguments{
\item{x}{numeric vector of length \eqn{n}}

\item{family}{character \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}
}
\value{
Returns a numeric vector of length \eqn{n}.
}
\description{
Transform the linear predictor to predicted values/probabilities.
}
\examples{
x <- rnorm(10)
corila:::.mean_function(x, family = "binomial")
corila:::.mean_function(x, family = "poisson")

}
\keyword{internal}
