% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.residuals}
\alias{.residuals}
\title{Deviance Residuals}
\usage{
.residuals(y_obs, y_fit, family)
}
\arguments{
\item{y_obs}{\eqn{n}-dimensional vector of observed values}

\item{y_fit}{\eqn{n}-dimensional vector of fitted values or probabilities}

\item{family}{character \code{"gaussian"}, \code{"binomial"}, or \code{"poisson"}}
}
\description{
Calculates the deviance residuals.
}
\details{
This function is called by \code{\link[=residuals.cv.corila]{residuals.cv.corila()}}.
}
\examples{
n <- 10
y_obs <- stats::rbinom(n = n, size = 1, prob = 0.2)
y_fit <- stats::runif(n = n)
corila:::.residuals(y_obs = y_obs, y_fit = y_fit, family = "binomial")

}
\keyword{internal}
