% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.set_candidates}
\alias{.set_candidates}
\title{Candidate values}
\usage{
.set_candidates(tune)
}
\arguments{
\item{tune}{character string for determining the candidate values
for the hyperparameters:
\itemize{
\item "none":
fixed weights and exponents
(\code{wgt_local}=1, \code{exp_local}=1, \code{wgt_global}=0),
no tuning
\item "weight":
fixed exponents (\code{exp_local}=0, \code{exp_global}=1),
tuning \code{wgt_local}=1-\code{wgt_global}
\item "exponent":
fixed weights (\code{wgt_local}=1, \code{wgt_global}=0),
tuning \code{exp_local}
\item "bivariate":
tuning \code{wgt_local}=1-\code{wgt_global} and \code{exp_local}=\code{exp_global}
\item "factorial":
tuning \code{wgt_local}, \code{exp_local}, \code{wgt_global}, \code{exp_global}
}

(to implement: list with slots
\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global})}
}
\value{
Returns a data frame with
the slots \code{"wgt_local"} and \code{"exp_local"}
for the local prior information
and the slots \code{"wgt_global"} and \code{"exp_global"}
for the global prior information.
}
\description{
Sets candidate values for hyperparameters.
}
\examples{
corila:::.set_candidates(tune = "none")

}
\seealso{
This function is called by \code{\link[=cv.corila]{cv.corila()}}.
}
\keyword{internal}
