% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.set_candidates}
\alias{.set_candidates}
\title{Candidate Values}
\usage{
.set_candidates(tune)
}
\value{
Returns a data frame with
the slots "wgt_local" and "exp_local" for the local prior information
and the slots "wgt_global" and "exp_global" for the global prior information.
}
\description{
Set candidate values for hyperparameters.
}
\examples{
.set_candidates(tune = "none")

}
\keyword{internal}
