% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_effects}
\alias{.simulate_effects}
\title{Simulate effects}
\usage{
.simulate_effects(
  group,
  prob_group = 0.5,
  prob_predictor = 0.8,
  signal_strength = 1
)
}
\arguments{
\item{group}{integer vector (length \eqn{p}, minimum 1, maximum \eqn{p})}

\item{prob_group}{probability for each group to be active:
numeric scalar in the unit interval}

\item{prob_predictor}{probability for each predictor in an active group to be active:
numeric scalar in the unit interval}

\item{signal_strength}{non-negative numeric scalar (default: \code{signal_strength=1})
for multiplying the effect sizes
(to increase or decrease the signal strength)}
}
\value{
Returns a numeric vector of length \eqn{p}.
}
\description{
Simulates effect vector.
}
\examples{
\dontshow{.simulate_effects <- corila:::.simulate_effects}
.simulate_effects(group = rep(c(1L:5L), each = 3L))

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
