% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate_predictors}
\alias{.simulate_predictors}
\title{Simulate predictors}
\usage{
.simulate_predictors(n, p = NULL, group = NULL, rho = 0)
}
\arguments{
\item{n}{number of observations (positive integer)}

\item{p}{number of predictors (positive integer)}

\item{group}{integer vector (length \eqn{p}, minimum 1, maximum \eqn{p})}

\item{rho}{correlation coefficient for predictors within the same group:
numeric scalar in the unit interval (minimum 0, maximum 1)}
}
\value{
Returns a numeric matrix with \eqn{n} rows (observations)
and \eqn{p} columns (predictors).
}
\description{
Simulates predictor matrix.
}
\examples{
\dontshow{.simulate_predictors <- corila:::.simulate_predictors}
.simulate_predictors(n = 5L, p = 7L)

.simulate_predictors(n = 5L, group = rep(c(1L, 2L), each = 3L), rho = 1.0)

}
\seealso{
This function is called by \code{\link[=simulate_data]{simulate_data()}}.
}
\keyword{internal}
