% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.type}
\alias{.type}
\title{Name (helper function)}
\usage{
.type(alpha)
}
\value{
Returns a character string
("ridge regression", "lasso regression", "elastic net regression",
"multi-penalty ridge regression",
or "Pearson/Spearman/Kendall correlation")
}
\description{
Names the method used for obtaining initial or final coefficients.
}
\examples{
corila:::.type(alpha = 0)

}
\seealso{
This function is called by \code{\link{print.summary.cv.corila}()}.
}
\keyword{internal}
