% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{.validate}
\alias{.validate}
\title{Argument Check}
\usage{
.validate(
  x,
  y,
  group,
  include,
  family,
  hyper,
  alpha_init,
  alpha_final,
  cor,
  foldid,
  nfolds,
  lambda_init
)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector,
where \eqn{n_0} is the number of observations used for model training}

\item{group}{\strong{(i)} \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\strong{(ii)} list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\strong{(iii)} \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{include}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE}, "primary predictors")
or must be excluded from the final model
(\code{FALSE}, "auxiliary predictors")}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{alpha_init}{elastic net mixing parameter
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
for initial regression
(default: ridge penalisation with \code{alpha_init}=0);
alternative choices are
"pearson", "spearman", or "kendall"
to use initial correlation coefficients
(not implemented for \code{family="cox"}),
"multiridge" for multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
or \code{NA} to set all initial coefficients equal to 1}

\item{alpha_final}{elastic net mixing parameter for final regression
(default: lasso penalisation with \code{alpha_final}=1)}

\item{cor}{character string \code{"pearson"},
\code{"spearman"} (default),
or \code{"kendall"};
or \eqn{p \times p} correlation matrix}

\item{foldid}{\eqn{n}-dimensional vector containing the fold identifiers}

\item{nfolds}{integer specifying the number of folds}
}
\value{
Returns \code{NULL} or an error message.
}
\description{
Checks arguments of functions \code{\link[=corila]{corila()}} and \code{\link[=cv.corila]{cv.corila()}}.
}
\examples{
NA

}
\seealso{
This function repeatedly calls \code{\link[=.assert]{.assert()}}.
}
\keyword{internal}
