% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate}
\alias{.validate}
\title{Validate data}
\usage{
.validate(x, y, family)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector,
where \eqn{n_0} is the number of observations used for model training}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}
}
\value{
Returns \code{NULL} or an error message.
}
\description{
Validates the predictor matrix x and the outcome vector y.
}
\examples{
n <- 10
p <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
.validate(x = x, y = y, family = "gaussian")

}
\seealso{
\code{\link{.check}()}
}
\keyword{internal}
