% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.validate}
\alias{.validate}
\title{Argument check}
\usage{
.validate(
  na_action,
  x,
  y,
  group,
  primary,
  family,
  hyper,
  alpha_init,
  alpha_final,
  cor,
  foldid,
  nfolds,
  lambda_init,
  silent
)
}
\arguments{
\item{na_action}{character \code{"error"} to trigger an error
if any observation has a missing predictor or a missing response
or \code{"complete_cases"} to exclude observations
with a missing predictor or a missing response from model fitting
(while providing fitted values for these observations)}

\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{group}{group structure (three options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
}}

\item{primary}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE} for "primary predictors")
or must be excluded from the final model
(\code{FALSE} for "auxiliary predictors")}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{alpha_init}{elastic net mixing parameter
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
for initial regression
(default: ridge penalisation with \code{alpha_init}=0);
alternative choices are
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"}),
\code{"multiridge"} for multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
or \code{NA} to set all initial coefficients equal to 1}

\item{alpha_final}{elastic net mixing parameter for final regression
(default: lasso penalisation with \code{alpha_final}=1)}

\item{cor}{character string \code{"pearson"},
\code{"spearman"} (default),
or \code{"kendall"};
or \eqn{p \times p} correlation matrix}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers}

\item{nfolds}{integer specifying the number of folds}

\item{silent}{Should messages from \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
be suppressed? logical}
}
\value{
Returns a list with slots \code{n}, \code{p}, and \code{q} or an error message.
}
\description{
Checks arguments of functions \code{\link[=corila]{corila()}} and \code{\link[=cv.corila]{cv.corila()}}.
}
\details{
This function is called by \code{\link[=corila]{corila()}} and \code{\link[=cv.corila]{cv.corila()}}.
It repeatedly calls \code{\link[=.assert]{.assert()}}.
}
\seealso{
Use \code{\link[=.assert]{.assert()}} to validate individual arguments.
}
\keyword{internal}
