% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.estim_initial_coefs}
\alias{.estim_initial_coefs}
\alias{.estim_multiridge_coefs}
\alias{.estim_cor_coefs}
\alias{.estim_glmnet_coefs}
\title{Initial coefficients}
\usage{
.estim_initial_coefs(
  x,
  y,
  family = "gaussian",
  alpha_init = 0,
  group = NULL,
  foldid = NULL,
  nfolds = 10L,
  lambda = NULL,
  silent = FALSE
)

.estim_multiridge_coefs(x, y, group, family, foldid, nfolds, lambda)

.estim_cor_coefs(x, y, method)

.estim_glmnet_coefs(x, y, family, alpha, foldid, nfolds, lambda, silent)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{alpha_init}{A scalar specifying the method used for obtaining initial coefficients:
\itemize{
\item a numeric scalar in the unit interval
(\eqn{0 \leq} \code{alpha_init} \eqn{\leq 1})
to define the mixing parameter for elastic net regression
(default: ridge penalisation with \code{alpha_init}=0);
\item the character scalar \code{"pearson"}, \code{"spearman"}, or \code{"kendall"}
to use initial correlation coefficients
(not implemented for \code{family="cox"})
\item the character scalar\code{"multiridge"}
to use multi-penalty ridge regression
with one penalty for each group
(not implemented for \code{family="poisson"} or overlapping groups),
\item \code{NA} to set all initial coefficients equal to 1
}}

\item{group}{\eqn{p}-dimensional integer vector with entries in \eqn{\{1, \ldots, q\}}}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers
(minimum \eqn{1}, maximum \code{nfolds})}

\item{nfolds}{positive integer specifying the number of folds
(minimum \eqn{3}, maximum \eqn{n})

NB: If \code{foldid} is provided, \code{nfolds} is overwritten by \code{max(foldid)}.}

\item{lambda}{numeric scalar, or \code{NULL}
(determined by cross-validation)}

\item{silent}{Should messages from \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
be suppressed? (logical scalar, \code{FALSE} or \code{TRUE})}
}
\value{
Returns a list with two slots:
\itemize{
\item \code{coef}: numeric vector of length \eqn{p}
(estimated coefficients)
\item \code{lambda}: non-negative numeric scalar
(optimised regularisation parameter) or \code{NULL}
}
}
\description{
Estimate initial coefficients.
}
\details{
This function is called by \code{\link[=corila]{corila()}}.
It calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} or \code{\link[glmnet:glmnet]{glmnet::glmnet()}}
for an initial lasso, ridge, or elastic net regression,
\code{\link[=multiridge]{multiridge()}} for an initial multi-penalty ridge regression,
or \code{\link[stats:cor]{stats::cor()}} for initial correlation coefficients.
}
\examples{
\dontshow{.estim_initial_coefs <- corila:::.estim_initial_coefs}
# simulate data
set.seed(1)
n <- 50L
p <- 10L
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
beta <- rbinom(n = p, size = 1L, prob = 0.5) * rnorm(p)
y <- drop(x \%*\% beta)

# initial correlation coefficients
.estim_initial_coefs(x = x,
                     y = y,
                     family = "gaussian",
                     alpha_init = "spearman",
                     group = NULL,
                     foldid = NULL,
                     nfolds = 10L,
                     lambda = NULL)

# initial regression coefficients (cross-validating lambda)
foldid <- sample(seq_len(10L), size = n, replace = TRUE)
.estim_initial_coefs(x = x,
                     y = y,
                     family = "gaussian",
                     alpha_init = 0.0,
                     group = NULL,
                     foldid = foldid,
                     nfolds = 10L,
                     lambda = NULL,
                     silent = TRUE)

# initial regression coefficients (using fixed lambda)
.estim_initial_coefs(x = x,
                     y = y,
                     family = "gaussian",
                     alpha_init = 0.0,
                     group = NULL,
                     foldid = NULL,
                     nfolds = 10L,
                     lambda = 0.2)

}
\keyword{internal}
