% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{fitted.cv.corila}
\alias{fitted.cv.corila}
\title{Fitted values}
\usage{
\method{fitted}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\description{
Extracts fitted values.
}
