% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{fitted.cv.corila}
\alias{fitted.cv.corila}
\title{Fitted values}
\usage{
\method{fitted}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns a numeric vector of length \eqn{n_0}
(one entry for each training observation).
}
\description{
Extracts fitted values.
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# fitting the model
n <- 10; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)

# using S3 methods
coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
\seealso{
Use \link[=predict.cv.corila]{predict()} to obtain predicted values
(i.e., for testing observations).
}
