% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forescale}
\alias{forescale}
\title{Standardisation}
\usage{
forescale(x, y = NULL, family = NULL, pars = NULL)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector or \code{NULL}, only for Gaussian family}

\item{family}{character string \code{"gaussian"}, \code{NULL}}

\item{pars}{list or \code{NULL}}
}
\value{
- standardised \eqn{n_0 \times p} predictor matrix \eqn{x}
- standardised \eqn{n_0}-dimensional response vector \eqn{y} (only if \code{family="gaussian"} or \code{pars$family="gaussian"}; otherwise output equals input)
- list \code{pars} with slots \code{mu.x} and \code{sd.x} (\eqn{p}-dimensional vectors of means and standard deviations of the predictor variables), and \code{mu.y} and \code{sd.y} (mean and standard deviation of response variable for Gaussian family, 0 and 1 for other families)
- character string \code{family} indicates the model (\code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}), determined by argument \code{family} or \code{pars$family}
}
\description{
Transforming variables to mean 0 and variance 1.
}
