% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forescale}
\alias{forescale}
\title{Standardisation}
\usage{
forescale(x, y = NULL, family = NULL, pars = NULL)
}
\arguments{
\item{x}{\eqn{n_0 \times p} predictor matrix, where \eqn{n_0} is the number of observations used for model training and \eqn{p} is the number of variables}

\item{y}{\eqn{n_0}-dimensional response vector or \code{NULL}, only for Gaussian family}

\item{family}{character string \code{"gaussian"}, \code{NULL}}

\item{pars}{list or \code{NULL} (see section "Value")}
}
\value{
\itemize{
\item standardised \eqn{n_0 \times p} predictor matrix \eqn{x}
\item standardised \eqn{n_0}-dimensional response vector \eqn{y} (only if \code{family="gaussian"} or \code{pars$family="gaussian"}; otherwise output equals input)
\item list \code{pars} with slots \code{mu.x} and \code{sd.x} (\eqn{p}-dimensional vectors of means and standard deviations of the predictor variables), and \code{mu.y} and \code{sd.y} (mean and standard deviation of response variable for Gaussian family, 0 and 1 for other families)
\item character string \code{family} indicates the model (\code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}), determined by argument \code{family} or \code{pars$family}
}
}
\description{
Transforming variables to mean 0 and variance 1.
}
\examples{

# simulate data
n <- 100; p <- 3
sd <- stats::rpois(n=p,lambda=5)
x <- sapply(X=sd,FUN=function(x) stats::rnorm(n=n,sd=x))
beta <- stats::rnorm(n=p)
y <- x \%*\% beta + stats::rnorm(n=n)

# without standardisation
lm1 <- stats::lm(y~x)
y_hat1 <- fitted(lm1)
coef1 <- coef(lm1)

# with standardisation
scale <- forescale(x=x,y=y,family="gaussian")
lm2 <- stats::lm(scale$y~scale$x)
result <- backscale(pars=scale$pars,y=fitted(lm2),coef=coef(lm2))
y_hat2 <- result$y_original
coef2 <- result$coef

# equality
all.equal(y_hat1,y_hat2)
all.equal(coef1,coef2,check.attributes=FALSE)

}
\seealso{
Use function \code{\link{backscale}} to bring coefficients and predictions back to original scale.
}
