% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{holdout}
\alias{holdout}
\title{Comparison with hold-out}
\usage{
holdout(
  x_train,
  y_train,
  group,
  type,
  include,
  family,
  alpha.init = 0,
  alpha.final = 1,
  x_test = NULL,
  y_test = NULL,
  nfolds = 10,
  foldid = NULL,
  method = NULL,
  seed = NULL,
  init.multi = FALSE,
  trial = TRUE,
  tune = "both"
)
}
\arguments{
\item{x_train}{\eqn{n_0 \times p} matrix}

\item{y_train}{\eqn{n_0}-dimensional vector}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices (in \eqn{\{1,\ldots,q\}}) or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices (in \eqn{\{1,\ldots,p\}}) or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{th}}} row and the \eqn{k^{\text{th}}} column indicates whether information should be transferred from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable}

\item{type}{\eqn{p}-dimensional vector indicating the modalities, or \code{NULL} (assuming that all variables are from the same modality)}

\item{include}{\eqn{p}-dimensional logical vector indicating whether a predictor may be included in (\code{TRUE}) or must be excluded from (\code{FALSE}) the final model}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{alpha.init}{elastic net mixing parameter for initial regression (default: ridge penalisation with \code{alpha.init}=0)}

\item{alpha.final}{elastic net mixing parameter for final regression (default: lasso penalisation with \code{alpha.final}=1)}

\item{x_test}{\eqn{n_1 \times p} matrix}

\item{y_test}{\eqn{n_1}-dimensional vector}

\item{nfolds}{number of internal cross-validation folds (integer scalar)}

\item{foldid}{internal cross-validation fold identifiers (\eqn{p}-dimensional integer vector)}

\item{method}{character vector listing all methods to be compared}

\item{seed}{random seed (integer scalar) for reproducibility, or \code{NULL}}

\item{init.multi}{Use multi-penalty ridge regression (one penalty for each group) to obtain initial coefficients (\code{TRUE} or \code{FALSE})? Not implemented for \code{family="poisson"}.}

\item{trial}{logical}

\item{tune}{character \code{"wgt"}, \code{"exp"}, or \code{"both"} for determining the candidate values for the hyperparameters; or list with slots \code{wgt.local}, \code{wgt.global}, \code{exp.local}, and \code{exp.global} (not yet implemented)}
}
\value{
Returns a list with the following slots:
\itemize{
\item \eqn{n_1}-dimensional vector \code{y_hat} containing predicted values
\item \eqn{p}-dimensional vector \code{coef} containing estimated coefficients
\item numerical vector \code{difftime} indicating the computation time of each \code{method}
}
}
\description{
Compares methods using hold-out method
}
\examples{
\donttest{
data <- simulate()
results <- holdout(x_train=data$x_train,y_train=data$y_train,group=data$group,type=data$type,include=rep(c(TRUE,FALSE),each=80),x_test=data$x_test,y_test=data$y_test,family=data$info$family,method=c("mean","ridge","lasso","corila")) # Why does holdout require y_test? Try to remove this
}
}
