% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{holdout}
\alias{holdout}
\title{Comparison with hold-out}
\usage{
holdout(
  x_train,
  y_train,
  group,
  type,
  family,
  x_test = NULL,
  y_test = NULL,
  nfolds = 10,
  foldid = NULL,
  method = NULL,
  seed = NULL,
  init.multi = FALSE,
  trial = TRUE
)
}
\arguments{
\item{x_train}{\eqn{n_0 \times p} matrix}

\item{y_train}{\eqn{n_0}-dimensional vector}

\item{group}{\emph{(i)} \eqn{p}-dimensional vector of group indices or labels, \emph{(ii)} list with \eqn{q} slots containing the variable indices or labels, or \emph{(iii)} \eqn{p \times p} matrix, where the entry in the \eqn{j^{\text{(th)}}} row and the \eqn{k^{\text{(th)}}} column indicates whether information should be transferred from the \eqn{j^{\text{(th)}}} to the \eqn{k^{\text{(th)}}} variable}

\item{type}{\eqn{p}-dimensional vector}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"} or \code{"cox"}}

\item{x_test}{\eqn{n_1 \times p} matrix}

\item{y_test}{\eqn{n_1}-dimensional vector}

\item{nfolds}{number of internal cross-validation folds (integer scalar)}

\item{foldid}{internal cross-validation fold identifiers (\eqn{p}-dimensional integer vector)}

\item{method}{character vector listing all methods to be compared}

\item{seed}{random seed (integer scalar)}

\item{init.multi}{logical}

\item{trial}{logical}
}
\value{
Returns a list with the following slots:
\itemize{
\item \eqn{n_1}-dimensional vector \code{y_hat} containing predicted values
\item \eqn{p}-dimensional vector \code{coef} containing estimated coefficients
\item numerical vector \code{difftime} indicating the computation time of each \code{method}
}
}
\description{
Compares methods using hold-out method
}
\examples{
NULL

}
