% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{methods}
\alias{methods}
\title{List of methods for class \code{"cv.corila"}}
\value{
\link[=coef.cv.corila]{coef()} returns a \eqn{(1 +) p}-dimensional vector,
\link[=predict.cv.corila]{predict()} returns an \eqn{n_1}-dimensional vector,
\link[=fitted.cv.corila]{fitted()} and \link[=residuals.cv.corila]{residuals()}
return an \eqn{n_0}-dimensional vector.
See individual methods for details.
}
\description{
Implemented S3 methods for objects of class \code{"cv.corila"}:
\itemize{
\item \link[=coef.cv.corila]{coef()}:
extracts estimated coefficients
\item \link[=predict.cv.corila]{predict()}:
calculates predicted values
\item \link[=fitted.cv.corila]{fitted()}:
extracts fitted values
\item \link[=residuals.cv.corila]{residuals()}:
calculates deviance residuals
\item \link[=plot.cv.corila]{plot()}:
visualises observed vs fitted values and estimated coefficients
\item \link[=print.cv.corila]{print()}:
prints information to the console
\item \link[=summary.cv.corila]{summary()}:
summarises the fitted model
}
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# fitting the model
n <- 10; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)

# using S3 methods
coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
\seealso{
Use \code{\link[=cv.corila]{cv.corila()}} to fit the model.
}
