% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ridge.R
\name{multiridge}
\alias{multiridge}
\title{Multi-penalty ridge regression}
\usage{
multiridge(
  x,
  y,
  group,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10L,
  penalties = NULL
)
}
\arguments{
\item{x}{predictors:
\eqn{n \times p} numeric matrix}

\item{y}{response:
\eqn{n}-dimensional vector}

\item{group}{\eqn{p}-dimensional integer vector with entries in \eqn{\{1, \ldots, q\}}}

\item{family}{character \code{"gaussian"} (not \code{"linear"}),
\code{"binomial"} (not \code{"logistic"}), or \code{"cox"}}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers
(minimum \eqn{1}, maximum \code{nfolds})}

\item{nfolds}{positive integer specifying the number of folds
(minimum \eqn{3}, maximum \eqn{n})

NB: If \code{foldid} is provided, \code{nfolds} is overwritten by \code{max(foldid)}.}

\item{penalties}{\eqn{q}-dimensional vector of non-negative penalty parameters,
or \code{NULL} (cross-validation)}
}
\value{
Returns an object of class \code{"multiridge"},
a list with the following slots:
\itemize{
\item slots from \link[multiridge:IWLSridge]{IWLSridge()} or
\link[multiridge:IWLSCoxridge]{IWLSCoxridge()}
\item character \code{family}
with value \code{"gaussian"} (also for \code{"linear"}),
\code{"binomial"} (also for \code{"logistic"}), or \code{"cox"}
\item \eqn{q}-dimensional vector \code{penalties}
containing optimised regularisation hyperparameters
(one for each predictor group)
\item list \code{indices}
with \code{nfolds} slots (one for each cross-validation fold),
each containing the indices of the observations
\item list \code{datablocks}
with \eqn{q} slots (one for each predictor group),
each containing an \eqn{n_0 \times p_k} matrix,
where \eqn{k \in \{1, \ldots, q\}}
\item \eqn{p}-dimensional group vector \code{group} (see argument)
\item list \code{pars} with slots \code{family} (see above),
the \eqn{p}-dimensional vectors \code{mu.x} and \code{sd.x}
and the scalars \code{mu.y} and \code{sd.y}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Fits multi-penalty ridge regression
(tuning regularisation hyperparameters
and estimating regression coefficients).
This is a wrapper function of some functions
from the \link[multiridge:multiridge-package]{multiridge-package}.
}
\details{
The numbers of observations (samples) for training or testing
are indicated by \eqn{n_0} and \eqn{n_1}, respectively,
the number of predictors (features) is indicated by \eqn{p},
and the number of predictor group is indicated by \eqn{q}.
Observations are indexed by \eqn{i} in \eqn{\{1, \ldots, n\}},
predictors are indexed by \eqn{j} in \eqn{\{1, \ldots, p\}},
and predictor groups are indexed by \eqn{k} in \eqn{\{1, \ldots, q\}}.
The number of predictors in the \eqn{k^{\text{th}}} group
is indicated by \eqn{p_k},
with \eqn{\sum_{k=1}^q p_k = p} for non-overlapping groups.
}
\examples{
\donttest{
data <- simulate_data(prob_primary = 1.0)
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group)
beta_hat <- coef(model)
y_hat <- predict(object = model, newx = data$x_test)
}

# examples for automatic mutation testing (with the R package autotest)
data <- simulate_data(prob_primary = 1.0)
foldid <- sample(seq_len(10L), size = nrow(data$x_train), replace = TRUE)
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group,
                    foldid = foldid)
penalties <- abs(rnorm(length(unique(data$group))))
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group,
                    penalties = penalties)

}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee}
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\doi{10.1080/10618600.2021.1904962}.
}
\seealso{
Extract coefficients with \link[=coef.multiridge]{coef()}
or make predictions with \link[=predict.multiridge]{predict()}.
Use \code{\link[=cv.corila]{cv.corila()}} to estimate sparse models.

This wrapper function calls various functions from the
\link[multiridge:multiridge-package]{multiridge-package},
namely
\link[multiridge:createXXblocks]{createXXblocks()},
\link[multiridge:fastCV2]{fastCV2()},
\link[multiridge:CVfolds]{CVfolds()},
\link[multiridge:optLambdasWrap]{optLambdasWrap()},
\link[multiridge:SigmaFromBlocks]{SigmaFromBlocks()},
\link[multiridge:IWLSridge]{IWLSridge()}, and
\link[multiridge:IWLSCoxridge]{IWLSCoxridge()}.

The \link[multiridge:multiridge-package]{multiridge-package}
accepts not only
an \eqn{n \times p} matrix but also
a list of length \eqn{q} of \eqn{n \times p_k} matrices,
with \eqn{k} in \eqn{\{1, \ldots, q\}}.
}
\keyword{classif}
\keyword{methods}
\keyword{models}
\keyword{regression}
