% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiridge}
\alias{multiridge}
\title{Multi-Penalty Ridge Regression}
\usage{
multiridge(x, y, z, family, penalties = NULL)
}
\arguments{
\item{x, y, z, family, penalties}{to be documented}
}
\value{
Returns an object of class \code{multiridge}.
}
\description{
Fits multi-penalty ridge regression.
}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee},
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast Cross-validation for Multi-penalty High-dimensional Ridge Regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847
\href{https://doi.org/10.1080/10618600.2021.1904962}{doi: 10.1080/10618600.2021.1904962}.
}
