% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiridge}
\alias{multiridge}
\title{Multi-Penalty Ridge Regression}
\usage{
multiridge(x, y, z, family, penalties = NULL)
}
\arguments{
\item{x, y, z, family, penalties}{to be documented}
}
\value{
Returns an object of class \code{multiridge}.
}
\description{
Fits multi-penalty ridge regression.
}
