% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-ridge.R
\name{multiridge}
\alias{multiridge}
\title{Multi-penalty ridge regression}
\usage{
multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10,
  penalties = NULL
)
}
\arguments{
\item{x}{predictors:
\eqn{n \times p} matrix}

\item{y}{response:
\eqn{n}-dimensional vector}

\item{z}{\eqn{p}-dimensional vector with entries in \eqn{\{1, \ldots, q\}}}

\item{family}{character \code{"linear"} (or \code{"gaussian"}),
\code{"logistic"} (or \code{"binomial"}), or \code{"cox"}}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers}

\item{nfolds}{integer specifying the number of folds}

\item{penalties}{\eqn{q}-dimensional vector of penalty parameters,
or \code{NULL} (cross-validation)}
}
\value{
Returns an object of class \code{"multiridge"},
a list with the following slots:
\itemize{
\item slots from \link[multiridge:IWLSridge]{IWLSridge()} or
\link[multiridge:IWLSCoxridge]{IWLSCoxridge()}
\item character \code{family}
with value \code{"gaussian"} (also for \code{"linear"}),
\code{"binomial"} (also for \code{"logistic"}), or \code{"cox"}
\item \eqn{q}-dimensional vector \code{penalties}
containing optimised regularisation hyperparameters
(one for each predictor group)
\item list \code{indices}
with \code{nfolds} slots (one for each cross-validation fold),
each containing the indices of the observations
\item list \code{datablocks}
with \eqn{q} slots (one for each predictor group),
each containing an \eqn{n_0 \times p_k} matrix,
where \eqn{k \in \{1, \ldots, q\}}
\item \eqn{p}-dimensional group vector \code{z} (see argument)
\item list \code{pars} with slots \code{family} (see above),
the \eqn{p}-dimensional vectors \code{mu.x} and \code{sd.x}
and the scalars \code{mu.y} and \code{sd.y}
}
}
\description{
Fits multi-penalty ridge regression
(tuning regularisation hyperparameters
and estimating regression coefficients).
This is a wrapper function of some functions
from the \link[multiridge:multiridge-package]{multiridge-package}.
}
\details{
The number of observations (samples) for training or testing
are indicated by \eqn{n_0} and \eqn{n_1}, respectively,
the number of variables (features) is indicated by \eqn{p},
and the number of variable groups is indicated by \eqn{q}.
Observations (samples) are indexed by \eqn{i} in \eqn{\{1, \ldots, n\}},
variables (features) are indexed by \eqn{j} in \eqn{\{1, \ldots, p\}},
and variable groups are indexed by \eqn{k} in \eqn{\{1, \ldots, q\}}.
The number of variables in the \eqn{k^{\text{th}}} group
is indicated by \eqn{p_k},
with \eqn{\sum_{k=1}^q p_k = p} for non-overlapping groups.
}
\examples{
# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)

\donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x \%*\% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
sapply(coef, function(x) mean((beta-x[-1])^2))

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
}

}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee}
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\href{https://doi.org/10.1080/10618600.2021.1904962}{doi: 10.1080/10618600.2021.1904962}. # nolint: line_length_linter.
}
\seealso{
Extract coefficients with \link[=coef.multiridge]{coef()}
or make predictions with \link[=predict.multiridge]{predict()}.
Use \code{\link[=cv.corila]{cv.corila()}} to estimate sparse models.

This wrapper function calls various functions from the
\link[multiridge:multiridge-package]{multiridge-package},
namely
\link[multiridge:createXXblocks]{createXXblocks()},
\link[multiridge:fastCV2]{fastCV2()},
\link[multiridge:CVfolds]{CVfolds()},
\link[multiridge:optLambdasWrap]{optLambdasWrap()},
\link[multiridge:SigmaFromBlocks]{SigmaFromBlocks()},
\link[multiridge:IWLSridge]{IWLSridge()}, and
\link[multiridge:IWLSCoxridge]{IWLSCoxridge()}.

The \link[multiridge:multiridge-package]{multiridge-package}
accepts not only
an \eqn{n \times p} matrix but also
a list of length \eqn{q} of \eqn{n \times p_k} matrices,
with \eqn{k} in \eqn{\{1, \ldots, q\}}.
}
\keyword{classif}
\keyword{methods}
\keyword{models}
\keyword{regression}
