% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiridge}
\alias{multiridge}
\title{Multi-Penalty Ridge Regression}
\usage{
multiridge(x, y, z, family, penalties = NULL)
}
\arguments{
\item{x}{predictors: \eqn{n \times p} matrix, or list of length \eqn{q} of \eqn{n \times p_k} matrices, with \eqn{k} in \eqn{\{1,\ldots,q\}}.}

\item{y}{response: \eqn{n}-dimensional vector}

\item{z}{groups: \eqn{p}-dimensional vector with entries in \eqn{\{1,\ldots,q\}} (if \code{x} is a matrix), or \code{NULL} (if \code{x} is a list of matrices)}

\item{family}{character \code{"linear"} (or \code{"gaussian"}), \code{"logistic"} (or \code{"binomial"}), or \code{"cox"}}

\item{penalties}{\eqn{q}-dimensional vector of penalty parameters, or \code{NULL} (cross-validation)}
}
\value{
Returns an object of class \code{multiridge}.
}
\description{
Fits multi-penalty ridge regression (tuning regularisation parameters and estimating regression coefficients).
}
\examples{

set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100,50)
z <- rep(x=seq_along(p),times=p)
x <- sapply(X=z,FUN=function(x) stats::rnorm(n=n,sd=x))
beta <- stats::rnorm(n=sum(p),mean=1,sd=0)*stats::rbinom(n=sum(p),size=1,prob=0.2)
eta <- x \%*\% beta
family <- "gaussian"
if(family=="gaussian"){
 y <- eta + 0.5*stats::rnorm(n=n,sd=stats::sd(eta))
} else if(family=="binomial"){
 y <- stats::rbinom(n=n,size=1,prob=1/(1+exp(-eta)))
} else if(family=="cox"){
 time <- stats::rexp(n=n,rate=exp(eta))
 status <- stats::rbinom(n=n,prob=0.5,size=1)
 y <- survival::Surv(time=time,event=status)
}
cond <- rep(x=c(TRUE,FALSE),times=c(n0,n1))

y_hat <- list()
object <- multiridge(x=x[cond,],y=y[cond],z=z,family=family)
y_hat$multiridge <- stats::predict(object,newx=x[!cond,])

# comparison
glmnet <- glmnet::cv.glmnet(x=x[cond,],y=y[cond],family=family,alpha=0)
y_hat$glmnet <- stats::predict(object=glmnet,newx=x[!cond,],type="response")

if(family=="gaussian"){
 metric <- sapply(X=y_hat,FUN=function(x) mean((x-y[!cond])^2))
} else if(family=="binomial"){
 metric <- sapply(X=y_hat,FUN=function(x) pROC::auc(response=y[!cond],predictor=as.vector(x),levels=c(0,1),direction="<"))
} else if(family=="cox"){
 metric <- sapply(X=y_hat,FUN=function(x) survival::concordance(y[!cond]~I(-x))$concordance)
}
metric

}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee},
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\href{https://doi.org/10.1080/10618600.2021.1904962}{doi: 10.1080/10618600.2021.1904962}.
}
\seealso{
Extract coefficients with \code{\link[=coef.multiridge]{coef}()} or make predictions with \code{\link[=predict.multiridge]{predict}()}.
}
