% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{nobs.cv.corila}
\alias{nobs.cv.corila}
\title{Observation Count}
\usage{
\method{nobs}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns a positive integer.
}
\description{
Extracts the number of observations.
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# fitting the model
n <- 10; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)

# using S3 methods
coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
