% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{plot.cv.corila}
\alias{plot.cv.corila}
\title{Plot Sparse Group Lasso (S3 method)}
\usage{
\method{plot}{cv.corila}(x, ...)
}
\arguments{
\item{x}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns NULL (invisible).
}
\description{
Plot method for class \code{"cv.corila"}.
}
\examples{
n <- 12 # decrease to 10 to check LOOCV
p <- 20
q <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
include <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, include = include)
print(object)
summary(object)
plot(object)

}
