% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cv.corila}
\alias{plot.cv.corila}
\title{Plot Sparse Group Lasso (S3 method)}
\usage{
\method{plot}{cv.corila}(x, ...)
}
\arguments{
\item{x}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Plot method for class \code{"cv.corila"}.
}
\details{
This function generates two figures:
\itemize{
\item a scatter plot of fitted versus observed values
for the Gaussian and the Poisson families,
a box plot of predicted probabilities for the two classes
for the binomial family,
or a histogram of fitted relative risks for the Cox model
\item estimated coefficients versus indices of predictors
}
}
\examples{
n <- 12 # decrease to 10 to check LOOCV
p <- 20
q <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)
print(object)
summary(object)

}
