% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{predict.corila}
\alias{predict.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{corila}(object, newx, index, s, ...)
}
\arguments{
\item{object}{object of class \code{"corila"}}

\item{newx}{\eqn{n_0 \times p} predictor matrix (training data)
to obtain fitted values,
\eqn{n_1 \times p} predictor matrix (testing data)
to obtain predicted values}

\item{index}{integer scalar specifying the index of the mixing hyperparameter(s)}

\item{s}{numeric vector specifying the values of the regularisation hyperparameter}

\item{...}{(for compatibility with \link[stats:predict]{stats::predict})}
}
\value{
Returns fitted or predicted values in an
\eqn{n_0 \times m}-dimensional or
\eqn{n_1 \times m}-dimensional matrix, respectively.
}
\description{
Makes prediction from an object of class \code{"corila"}.
}
\examples{
\donttest{
\dontshow{corila <- corila:::corila}
# simulation
n <- 100L
p <- 50L
group <- rep(x = seq_len(10L), each = 5L)
primary <- rep(x = TRUE, times = p)
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n = n)

# model fitting
hyper <- data.frame(wgt_local = 0.5, exp_local = 1.0,
                    wgt_global = 0.5, exp_global = 1.0)
object <- corila(x = x,
                 y = y,
                 group = group,
                 primary = primary,
                 family = "gaussian",
                 alpha_init = 0.0,
                 alpha_final = 1.0,
                 cor = "spearman",
                 foldid = NULL,
                 nfolds = 10L,
                 hyper = hyper,
                 threshold = 0.0,
                 lambda_init = NULL)

y_hat <- stats::predict(object, newx = x, index = 1L, s = 0.0)
}

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Estimate parameters with \code{\link[=corila]{corila()}},
or estimate parameters and tune hyperparameters
with \code{\link[=cv.corila]{cv.corila()}}.
}
\keyword{internal}
