% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.corila}
\alias{predict.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{corila}(object, newx, index, s, ...)
}
\arguments{
\item{object}{object of class \code{corila}}

\item{newx}{\eqn{n_1 \times p} matrix}

\item{index}{integer scalar specifying the index of the mixing hyperparameter(s)}

\item{s}{numeric scalar specifying the value of the regularisation hyperparameter}

\item{...}{(not used)}
}
\value{
Returns fitted or predicted values in an \eqn{n_0}-dimensional or \eqn{n_1}-dimensional vector, respectively.
}
\description{
Makes prediction from an object of class \code{corila}.
}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Estimate parameters with \code{\link{corila}()}, or estimate parameters and tune hyperparameters with \code{\link{cv.corila}()}.
}
