% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.corila}
\alias{predict.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{corila}(object, newx, index, s, ...)
}
\arguments{
\item{object}{object of class \code{"corila"}}

\item{newx}{\eqn{n_0 \times p} predictor matrix (training data) to obtain fitted values, \eqn{n_1 \times p} predictor matrix (testing data) to obtain predicted values}

\item{index}{integer scalar specifying the index of the mixing hyperparameter(s)}

\item{s}{numeric scalar specifying the value of the regularisation hyperparameter}

\item{...}{(not used)}
}
\value{
Returns fitted or predicted values in an \eqn{n_0}-dimensional or \eqn{n_1}-dimensional vector, respectively.
}
\description{
Makes prediction from an object of class \code{"corila"}.
}
\examples{
\donttest{
# simulation
n <- 100
p <- 50
group <- rep(x=1:10,each=5)
type <- rep(x=1,times=length(group))
include <- NULL
x <- matrix(data=rnorm(n*p),nrow=n,ncol=p)
y <- rnorm(n=n)

# model fitting
hyper <- data.frame(exp.local=1,wgt.local=0.5,exp.global=1,wgt.global=0.5)
object <- corila(x,y,group,include,type,family="gaussian",hyper=hyper)

y_hat <- stats::predict(object,newx=x,index=1,s=0)
}
}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Estimate parameters with \code{\link{corila}()}, or estimate parameters and tune hyperparameters with \code{\link{cv.corila}()}.
}
