% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.cv.corila}
\alias{predict.cv.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{cv.corila}(object, newx, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{newx}{\eqn{n_0 \times p} predictor matrix (training data)
to obtain fitted values,
\eqn{n_1 \times p} predictor matrix (testing data)
to obtain predicted values}

\item{s}{character \code{"lambda.min"} or numeric non-negative scalar}

\item{...}{(for compatibility with \link[stats:predict]{stats::predict})}
}
\value{
Returns fitted or predicted values in an
\eqn{n_0 \times m}-dimensional or
\eqn{n_1 \times m}-dimensional matrix, respectively.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Makes predictions from an object of class \code{"cv.corila"}.
}
\details{
This function calls
\code{\link[=.expand_auxiliary]{.expand_auxiliary()}} for handling auxiliary predictors,
\code{\link[=.forescale]{.forescale()}}
for standardising the predictor matrix,
and \code{\link[=.backscale]{.backscale()}}
for bringing predicted values back to the original scale
(if \code{family="gaussian"}).
}
\examples{
\donttest{
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = data$group,
                   primary = data$primary)
beta_hat <- coef(object = model)
y_hat <- predict(object = model, newx = data$x_test)
}

# example for automatic mutation testing (with the R package autotest)
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = as.double(data$group),
                   primary = data$primary,
                   alpha_init = 0.0,
                   threshold = 0.0,
                   foldid = rep(1:10, length.out = nrow(data$x_train)))

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link[=cv.corila]{cv.corila()}},
extract coefficients with \link[=coef.cv.corila]{coef()},
and extract fitted values with \link[=fitted.cv.corila]{fitted()}.
}
\keyword{methods}
