% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.cv.corila}
\alias{predict.cv.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{cv.corila}(object, newx, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class "cv.corila"}

\item{newx}{\eqn{n_1 \times p} matrix}

\item{s}{character "lambda.min" or numeric value}

\item{...}{(not used)}
}
\value{
Returns fitted or predicted values in an \eqn{n_0}-dimensional or \eqn{n_1}-dimensional vector, respectively.
}
\description{
Makes predictions from an object of class \code{cv.corila}.
}
\seealso{
\code{\link{cv.corila}()}, \code{\link{coef.cv.corila}()}
}
