% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.cv.corila}
\alias{predict.cv.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{cv.corila}(object, newx, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{newx}{\eqn{n_1 \times p} matrix}

\item{s}{character \code{"lambda.min"} or numeric value}

\item{...}{(not used)}
}
\value{
Returns fitted or predicted values in an \eqn{n_0}-dimensional or \eqn{n_1}-dimensional vector, respectively.
}
\description{
Makes predictions from an object of class \code{"cv.corila"}.
}
\examples{
\donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100,50)
z <- rep(x=seq_along(p),times=p)
x <- sapply(X=z,FUN=function(x) stats::rnorm(n=n,sd=x))
beta <- stats::rnorm(n=sum(p),mean=1,sd=0)*stats::rbinom(n=sum(p),size=1,prob=0.2)
eta <- x \%*\% beta
family <- "gaussian"
if(family=="gaussian"){
 y <- eta + 0.5*stats::rnorm(n=n,sd=stats::sd(eta))
} else if(family=="binomial"){
 y <- stats::rbinom(n=n,size=1,prob=1/(1+exp(-eta)))
} else if(family=="cox"){
 time <- stats::rexp(n=n,rate=exp(eta))
 status <- stats::rbinom(n=n,prob=0.5,size=1)
 y <- survival::Surv(time=time,event=status)
}
cond <- rep(x=c(TRUE,FALSE),times=c(n0,n1))

y_hat <- coef <- list()

# standard lasso regression
object <- glmnet::cv.glmnet(x=x[cond,],y=y[cond],family=family,alpha=1)
coef$glmnet <- stats::coef(object=object,s="lambda.min")
y_hat$glmnet <- stats::predict(object=object,newx=x[!cond,],type="response",s="lambda.min")

# flexible group lasso regression
object <- cv.corila(x=x[cond,],y=y[cond],group=z,family=family)
coef$corila <- stats::coef(object=object)
y_hat$corila <- stats::predict(object=object,newx=x[!cond,])

# estimation performance
sapply(coef,function(x) stats::cor(beta,x[-1]))
sapply(coef,function(x) mean((beta-x[-1])^2))

# predictive performance
if(family=="gaussian"){
 metric <- sapply(X=y_hat,FUN=function(x) mean((x-y[!cond])^2))
} else if(family=="binomial"){
 metric <- sapply(X=y_hat,FUN=function(x) pROC::auc(response=y[!cond],predictor=as.vector(x),levels=c(0,1),direction="<"))
} else if(family=="cox"){
 metric <- sapply(X=y_hat,FUN=function(x) survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
}
}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2025).
"A flexible version of the sparse group lasso".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link{cv.corila}()} and extract coefficients with \code{\link{coef.cv.corila}()}.
}
