% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.cv.corila}
\alias{predict.cv.corila}
\title{predict (S3 method)}
\usage{
\method{predict}{cv.corila}(object, newx, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{newx}{\eqn{n_0 \times p} predictor matrix (training data)
to obtain fitted values,
\eqn{n_1 \times p} predictor matrix (testing data)
to obtain predicted values}

\item{s}{character \code{"lambda.min"} or numeric value}

\item{...}{(for compatibility with \link[stats:predict]{stats::predict})}
}
\value{
Returns fitted or predicted values in an
\eqn{n_0 \times m}-dimensional or
\eqn{n_1 \times m}-dimensional matrix, respectively.
}
\description{
Makes predictions from an object of class \code{"cv.corila"}.
}
\details{
This function calls
\code{\link[=.expand_auxiliary]{.expand_auxiliary()}} for handling auxiliary predictors,
\code{\link[=.forescale]{.forescale()}}
for standardising the predictor matrix,
and \code{\link[=.backscale]{.backscale()}}
for bringing predicted values back to the original scale
(if \code{family="gaussian"}).
}
\examples{
data <- simulate_data()
model <- cv.corila(x = data$x_train,
                   y = data$y_train,
                   group = data$group,
                   primary = data$primary,
                   tune = "none")

#\dontshow{
#model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary,
#                    tune = "none")
# }
# \donttest{
# model <- cv.corila(x = data$x_train,
#                    y = data$y_train,
#                    group = data$group,
#                    primary = data$primary)
# }

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2026).
"Sparse modelling with grouped and correlated features
allowing for privileged information".
\emph{In preparation}.
}
\seealso{
Fit models with \code{\link[=cv.corila]{cv.corila()}},
extract coefficients with \link[=coef.cv.corila]{coef()},
and extract fitted values with \link[=fitted.cv.corila]{fitted()}.
}
\keyword{methods}
