% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ridge.R
\name{predict.multiridge}
\alias{predict.multiridge}
\title{Make predictions}
\usage{
\method{predict}{multiridge}(object, newx, ...)
}
\arguments{
\item{object}{object of type \code{"multiridge"}}

\item{newx}{\eqn{n_0 \times p} predictor matrix (training data)
to obtain fitted values,
\eqn{n_1 \times p} predictor matrix (testing data)
to obtain predicted values}

\item{...}{(for compatibility with \link[stats:predict]{stats::predict})}
}
\value{
Returns an \eqn{n_0}-dimensional vector of fitted values
or an \eqn{n_1}-dimensional vector of predicted values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Makes predictions from a multi-penalty ridge regression model.
}
\examples{
\donttest{
data <- simulate_data(prob_primary = 1.0)
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group)
beta_hat <- coef(model)
y_hat <- predict(object = model, newx = data$x_test)

\dontshow{
# examples for automatic mutation testing (with the R package autotest)
data <- simulate_data(prob_primary = 1.0)
foldid <- sample(seq_len(10L), size = nrow(data$x_train), replace = TRUE)
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group,
                    foldid = foldid)
penalties <- abs(rnorm(length(unique(data$group))))
model <- multiridge(x = data$x_train, y = data$y_train, group = data$group,
                    penalties = penalties)
}
}

}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee}
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\doi{10.1080/10618600.2021.1904962}.
}
\seealso{
Fit models with \code{\link[=multiridge]{multiridge()}}
and extract coefficients with \link[=coef.multiridge]{coef()}.
}
\keyword{methods}
