% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.multiridge}
\alias{predict.multiridge}
\title{Make Predictions}
\usage{
\method{predict}{multiridge}(object, newx, ...)
}
\arguments{
\item{object}{object of class \code{"multiridge"}}

\item{newx}{\eqn{n_1 \times p} matrix}

\item{...}{(not used)}
}
\description{
Makes predictions from a multi-penalty ridge regression model.
}
\seealso{
[multiridge()], [coef.multiridge()]
}
