% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.multiridge}
\alias{predict.multiridge}
\title{Make Predictions}
\usage{
\method{predict}{multiridge}(object, newx, ...)
}
\arguments{
\item{object}{object of class \code{"multiridge"}}

\item{newx}{\eqn{n_1 \times p} matrix}

\item{...}{(not used)}
}
\value{
Returns an \eqn{n_0}-dimensional vector of fitted values or an \eqn{n_1}-dimensional vector of predicted values.
}
\description{
Makes predictions from a multi-penalty ridge regression model.
}
\references{
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
\href{https://orcid.org/0000-0001-7715-1446}{Mirrelijn M. van Nee},
and
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger}
(2021).
"Fast cross-validation for multi-penalty high-dimensional ridge regression"
\emph{Journal of Computational and Graphical Statistics}
30(4):835-847.
\href{https://doi.org/10.1080/10618600.2021.1904962}{doi: 10.1080/10618600.2021.1904962}.
}
\seealso{
Fit models with \code{\link{multiridge}()} and extract coefficients with \code{\link{coef.multiridge}()}.
}
