% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{print.cv.corila}
\alias{print.cv.corila}
\title{print (S3 method)}
\usage{
\method{print}{cv.corila}(x, ...)
}
\arguments{
\item{x}{object of class \code{"cv.corila"}.}

\item{...}{(not used)}
}
\value{
Prints \code{"object of class 'cv.corila'"} to the console.
}
\description{
Print method for class \code{"cv.corila"}.
}
\examples{
n <- 12 # decrease to 10 to check LOOCV
p <- 20
q <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
include <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, include = include)
print(object)
summary(object)

}
\seealso{
\link[=summary.cv.corila]{summary()}
}
