% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.cv.corila}
\alias{print.cv.corila}
\title{print (S3 method)}
\usage{
\method{print}{cv.corila}(x, ...)
}
\arguments{
\item{x}{object of class \code{"cv.corila"}}

\item{...}{(not used)}
}
\description{
Print method for class \code{"cv.corila"}.
}
\seealso{
summary.cv.corila
}
