% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.cv.corila}
\alias{residuals.cv.corila}
\title{Residuals}
\usage{
\method{residuals}{cv.corila}(object, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(for compatibility with \link[stats:residuals]{stats::residuals})}
}
\value{
Returns a numeric vector of length \eqn{n_0}
(one entry for each training observation).
}
\description{
Residuals
}
\details{
This function extracts the observed and fitted values from the fitted model
and calls the internal function \code{\link[=.residuals]{.residuals()}} to calculate the residuals.
}
\examples{
# listing S3 methods
methods(class = "cv.corila")

# fitting the model
n <- 10L; p <- 20L; q <- 5L
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1L, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)

# using S3 methods
coef(object)
predict(object, newx = x)
fitted(object)
residuals(object)
plot(object)
print(object)
summary(object)

}
