% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{.set_candidates}
\alias{.set_candidates}
\title{Candidate values}
\usage{
.set_candidates(tune)
}
\arguments{
\item{tune}{character string for determining the candidate values
for the hyperparameters:
\itemize{
\item \code{"none"}:
fixed weights and exponents
(\code{wgt_local}=1, \code{exp_local}=1, \code{wgt_global}=0),
no tuning
(share information within groups, ignore correlations)
\item \code{"weight"} (default):
fixed exponents (\code{exp_local}=0, \code{exp_global}=1),
tuning \code{wgt_local}=1-\code{wgt_global}
(find a compromise between sharing information within groups
and between correlated predictors)
\item \code{"exponent"}:
fixed weights (\code{wgt_local}=1, \code{wgt_global}=0),
tuning \code{exp_local}
(share information between correlated predictors in the same group,
determine level of trust in correlation coefficients)
\item \code{"bivariate"}:
tuning \code{wgt_local}=1-\code{wgt_global} and \code{exp_local}=\code{exp_global}
(find a compromise between sharing information
between predictors in the same group and between all predictors,
determine level of trust in correlation coefficients)
\item \code{"factorial"}:
tuning \code{wgt_local}, \code{exp_local}, \code{wgt_global}, \code{exp_global}
(unrestricted information sharing
with weights possibly not summing to one
and possibly different exponents)
}

(NB: It is currently not possible to provide a data frame with columns
\code{wgt_local}, \code{exp_local}, \code{wgt_global}, and \code{exp_global}.)}
}
\value{
Returns a data frame with
the slots \code{"wgt_local"} and \code{"exp_local"}
for the local prior information
and the slots \code{"wgt_global"} and \code{"exp_global"}
for the global prior information.
}
\description{
Sets candidate values for hyperparameters.
}
\examples{
\dontshow{.set_candidates <- corila:::.set_candidates}
.set_candidates(tune = "none")

}
\seealso{
This function is called by \code{\link[=cv.corila]{cv.corila()}}.
}
\keyword{internal}
