% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sign_precision}
\alias{sign_precision}
\title{Precision for sign variable}
\usage{
sign_precision(truth, estim)
}
\arguments{
\item{truth}{integer vector with values in \eqn{\{-1, 0, 1\}}}

\item{estim}{integer vector of same length with values in \eqn{\{-1, 0, 1\}}}
}
\value{
Returns a scalar between 0 (minimum precision) and 1 (maximum precision),
or \code{NA} if all estimated signs equal 0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates precision for ternary variables with support \eqn{\{-1, 0, 1\}},
i.e., the proportion of positive or negative estimated signs
that match the true sign.
}
\examples{
\dontshow{sign_precision <- corila:::sign_precision}
truth <- sample(x = c(-1L, 0L, 1L), size = 10L, replace = TRUE)
estim <- sample(x = c(-1L, 0L, 1L), size = 10L, replace = TRUE)
sign_precision(truth = truth, estim = estim) # observed value
sign_precision(truth = truth, estim = -truth) # lower limit 0
sign_precision(truth = truth, estim = truth) # upper limit 1
sign_precision(truth = truth, estim = 0L * estim) # not defined

}
\keyword{internal}
