% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sign_recall}
\alias{sign_recall}
\title{Recall for sign variable}
\usage{
sign_recall(truth, estim)
}
\arguments{
\item{truth}{integer vector with values in \eqn{\{-1, 0, 1\}}}

\item{estim}{integer vector of same length with values in \eqn{\{-1, 0, 1\}}}
}
\value{
Returns a scalar between 0 (minimum recall) and 1 (maximum recall),
or \code{NA} if all true signs equal 0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates recall for ternary variables with support \eqn{\{-1, 0, 1\}},
i.e., the proportion of correctly identified positive and negative signs.
}
\examples{
\dontshow{sign_recall <- corila:::sign_recall}
truth <- sample(x = c(-1L, 0L, 1L), size = 10L, replace = TRUE)
estim <- sample(x = c(-1L, 0L, 1L), size = 10L, replace = TRUE)
sign_recall(truth = truth, estim = estim) # observed value
sign_recall(truth = truth, estim = 0L * estim) # lower limit 0
sign_recall(truth = truth, estim = truth) # upper limit 1
sign_recall(truth = 0L * truth, estim = estim) # not defined

}
