% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Data simulation}
\usage{
simulate(
  family = "gaussian",
  n0 = 100L,
  n1 = 10000L,
  n_group = 20L,
  n_type = 2L,
  size_group = c(5L, 3L),
  effect_size = c(1, 1),
  corfac_feature = 0.5,
  corfac_type = 0.5,
  corfac_group = 0.25,
  n_group_causal = 2L,
  prop_causal = 0.5,
  noise_factor = 1,
  plot = FALSE,
  trial = FALSE
)
}
\arguments{
\item{family}{character \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{n0}{number of training observations
(positive integer)}

\item{n1}{number of testing observations
(positive integer)}

\item{n_group}{number of variable groups
(positive integer)}

\item{n_type}{number of variable types
(positive integer)}

\item{size_group}{size of variable groups (per variable type):
integer vector of length \code{n_type}}

\item{effect_size}{effect sizes (per variable type):
numeric vector of length \code{n_type}}

\item{corfac_feature}{decrease of correlation if different variable:
scalar in unit interval}

\item{corfac_type}{decrease of correlation if different type:
scalar in unit interval}

\item{corfac_group}{decrease of correlation if different group:
scalar in unit interval}

\item{n_group_causal}{number of causal groups:
integer}

\item{prop_causal}{proportion of causal features within causal groups:
scalar in unit interval}

\item{noise_factor}{noise factor:
numeric scalar}

\item{plot}{Attempt to visualise effects of and correlation between variables?
(\code{TRUE} or \code{FALSE})}

\item{trial}{logical (groups of negatively correlated subgroups)}
}
\value{
Returns a list with the following slots:
\itemize{
\item \eqn{n_0 \times p} matrix \code{x_train}
\item \eqn{p}-dimensional vector \code{type}
\item \eqn{p}-dimensional vector \code{group}
\item \eqn{n_0}-dimensional vector \code{y_train}
\item \eqn{n_1 \times p} matrix \code{x_test}
\item \eqn{n_1}-dimensional vector \code{y_test}
\item \eqn{p}-dimensional vector \code{beta}
\item data frame \code{info} with entries
\eqn{n_0}, \eqn{n_1}, \eqn{p}, \code{n_type},
\code{n_group}, and \code{family}
}
}
\description{
Simulates data with grouped predictor variables.
}
\examples{
data <- corila:::simulate()
dims <- function(x) {
   if (is.matrix(x)||is.data.frame(x)) {
     paste(base::dim(x), collapse = " x ")
   } else {
     paste0(base::length(x))
   }
}
sapply(X = data, FUN = dims)

}
\keyword{internal}
