% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Data simulation}
\usage{
simulate(
  family = "gaussian",
  n0 = 100,
  n1 = 10000,
  n.group = 20,
  n.type = 2,
  size.group = c(5, 3),
  effect.size = c(1, 1),
  corfac.feature = 0.5,
  corfac.type = 0.5,
  corfac.group = 0.25,
  n.group.causal = 2,
  prop.causal = 0.5,
  noise.factor = 1,
  plot = TRUE
)
}
\arguments{
\item{family}{character \code{"gaussian"}, \code{"binomial"}, \code{"poisson"} or \code{"cox"}}

\item{n0}{number of training observations}

\item{n1}{number of testing observations}

\item{n.group}{number of variable groups}

\item{n.type}{number of variable types}

\item{size.group}{size of variable groups (per variable type)}

\item{effect.size}{effect sizes (per variable type)}

\item{corfac.feature}{decrease of correlation if different variable}

\item{corfac.type}{decrease of correlation if different type}

\item{corfac.group}{decrease of correlation if different group}

\item{n.group.causal}{number of causal groups}

\item{prop.causal}{proportion of causal features within causal groups}

\item{noise.factor}{noise factor}

\item{plot}{Attempt to visualise effects of and correlation between variables? (\code{TRUE} or \code{FALSE})}
}
\value{
Returns a list with the following slots:
\itemize{
\item \eqn{n_0 \times p} matrix \code{x_train}
\item \eqn{p}-dimensional vector \code{type}
\item \eqn{p}-dimensional vector \code{group}
\item \eqn{n_0}-dimensional vector \code{y_train}
\item \eqn{n_1 \times p} matrix \code{x_test}
\item \eqn{n_1}-dimensional vector \code{y_test}
\item \eqn{p}-dimensional vector \code{beta}
\item data frame \code{info} with entries \eqn{n_0}, \eqn{n_1}, \eqn{p}, \code{n.type}, \code{n.group}, and \code{family}
}
}
\description{
Simulates data with grouped predictor variables
}
\examples{
data <- simulate()
dims <- function(x){
  if(is.matrix(x)||is.data.frame(x)){
    paste(base::dim(x),collapse=" x ")
  } else {
    paste0(base::length(x))
  }
}
sapply(X=data,FUN=dims)

}
