% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\alias{simulate}
\title{Data simulation}
\usage{
simulate_data(
  n0 = 50L,
  n1 = 20L,
  p = 30L,
  q = 10L,
  family = "gaussian",
  rho = 0.5,
  prob_primary = 0.5,
  signal_strength = 1,
  prob_group = 0.5,
  prob_predictor = 0.8,
  seed = 1L
)
}
\arguments{
\item{n0}{number of training observations
(positive integer)}

\item{n1}{number of testing observations
(positive integer)}

\item{p}{number of predictors
(positive integer)}

\item{q}{number of predictor groups
(positive integer)}

\item{family}{character string \code{"gaussian"}, \code{"binomial"},
\code{"poisson"}, or \code{"cox"}}

\item{rho}{correlation coefficient for predictors within the same group:
numeric scalar in the unit interval (minimum 0, maximum 1)}

\item{prob_primary}{probability for each predictor to be primary (rather than auxiliary):
numeric scalar in the unit interval
(minimum 0 leads to auxiliary predictors only,
maximum 1 leads to primary predictors only)}

\item{signal_strength}{non-negative numeric scalar (default: \code{signal_strength=1})
for multiplying the effect sizes
(to increase or decrease the signal strength)}

\item{prob_group}{probability for each group to be active:
numeric scalar in the unit interval}

\item{prob_predictor}{probability for each predictor in an active group to be active:
numeric scalar in the unit interval}

\item{seed}{random seed for reproducibility:
integer scalar}
}
\value{
Returns a list with multiple slots:
\itemize{
\item \code{x_train}:
predictor matrix of the training observations
(\eqn{n_0} rows, \eqn{p} columns)
\item \code{y_train}:
response vector of the training observations
(length \eqn{n_0})
\item \code{group}:
integer vector indicating the group of the predictors
(length \eqn{p})
\item \code{primary}:
logical vector indicating
primary (\code{TRUE}) and auxiliary (\code{FALSE}) predictors
(length \eqn{p})
\item \code{beta}:
numeric vector of the effects of the predictors on the response
(length \eqn{p})
\item \code{x_test}:
\eqn{n_1 \times p} predictor matrix for the test observations
\item \code{y_test}:
response vector for the test observations of length \eqn{n_1}
}

Training and testing observations are named \code{train_} or \code{test_},
respectively, followed by a number indexing the observations
(e.g., \code{train_1} or \code{test_1}).

Primary and auxiliary predictors are named \code{pri_} or \code{aux_}, respectively,
followed by a number indexing the predictor groups, a point,
and a number indexing the predictors within this group
(e.g., \code{pri_1.1} or \code{aux_1.1}).
}
\description{
Simulates a predictor matrix, an effect vector and a response vector.
The simulated datasets can be used for modelling a response based on
grouped and correlated primary and auxiliary predictors.
}
\details{
\itemize{
\item Use the objects \code{x_train}, \code{y_train}, \code{group}, and \code{primary}
for model training.
Estimated coefficients can be compared with \code{beta}.
\item Use the object \code{x_test} for model testing.
Predicted values can be compared with \code{y_test}.
}
}
\examples{
\donttest{
data <- simulate_data(n0 = 100, n1 = 10000)

coef <- y_hat <- list()

# standard lasso regression
object <- glmnet::cv.glmnet(x = data$x_train[, data$primary],
y = data$y_train)
temp <- stats::coef(object = object, s = "lambda.min")[-1L]
coef$glmnet <- c(temp[1L], ifelse(primary, temp[-1L], 0.0))
y_hat$glmnet <- stats::predict(object = object,
                               newx = data$x_test[, data$primary],
                               type = "response",
                               s = "lambda.min")

# flexible group lasso regression
object <- cv.corila(x = data$x_train, y = data$y_train,
                    group = data$group, primary = data$primary)
coef$corila <- stats::coef(object = object)
y_hat$corila <- stats::predict(object = object,
                               newx = data$x_test[, data$primary])

# selection performance (precision: higher = better)
vapply(X = coef,
       FUN = function(x) {
         calc_sign_prec(truth = sign(data$beta), estim = sign(x[-1L]))
       },
       FUN.VALUE = numeric(1L))

# predictive performance (mean squared error: lower = better)
vapply(X = y_hat,
       FUN = function(x) mean((x - data$y_test)^2.0),
       FUN.VALUE = numeric(1L))
}

}
\seealso{
This function calls the internal functions \code{\link[=.simulate_predictors]{.simulate_predictors()}},
\code{\link[=.simulate_effects]{.simulate_effects()}}, and \code{\link[=.simulate_response]{.simulate_response()}} for simulating the
predictor matrix, the effect vector, or the response vector, respectively.
}
