% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lasso.R
\name{summary.cv.corila}
\alias{summary.cv.corila}
\alias{print.summary.cv.corila}
\title{Summarising Sparse Group Lasso (S3 method)}
\usage{
\method{summary}{cv.corila}(object, ...)

\method{print}{summary.cv.corila}(x, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(not used)}

\item{x}{object of class \code{"summary.cv.corila"}}
}
\value{
Returns an invisible list with multiple slots.
}
\description{
Summary method for class \code{"cv.corila"}.
}
\examples{
n <- 12 # decrease to 10 to check LOOCV
p <- 20
q <- 5
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
include <- as.logical(rbinom(n = p, size = 1, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, include = include)
print(object)
summary(object)

}
\seealso{
\code{\link[=print.cv.corila]{print.cv.corila()}}
}
\keyword{internal}
