% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cv.corila}
\alias{summary.cv.corila}
\alias{print.summary.cv.corila}
\title{Summarising sparse group lasso (S3 method)}
\usage{
\method{summary}{cv.corila}(object, ...)

\method{print}{summary.cv.corila}(x, ...)
}
\arguments{
\item{object}{object of class \code{"cv.corila"}}

\item{...}{(for compatibility with \link[base:summary]{base::summary})}

\item{x}{object of class \code{"summary.cv.corila"}}
}
\value{
Returns an invisible list with multiple slots.
}
\description{
Summary method for class \code{"cv.corila"}.
}
\details{
\code{\link[=print.summary.cv.corila]{print.summary.cv.corila()}} uses the output from \code{\link[=summary.cv.corila]{summary.cv.corila()}}
to print readable information to the console.
It calls the helper function \code{\link[=.type]{.type()}} to name
the methods used to estimate initial and final coefficients.
}
\examples{
n <- 12L # decrease to 10 to check LOOCV
p <- 20L
q <- 5L
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rnorm(n)
group <- rep(seq_len(q), length.out = p)
primary <- as.logical(rbinom(n = p, size = 1L, prob = 0.5))
object <- cv.corila(x = x, y = y, group = group, primary = primary)
print(object)
summary(object)

}
\seealso{
\code{\link[=print.cv.corila]{print.cv.corila()}}
}
\keyword{internal}
