% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{.validate_na_action}
\alias{.validate_family}
\alias{.validate_x}
\alias{.validate_y}
\alias{.validate_y_hat}
\alias{.validate_primary}
\alias{.validate_cor}
\alias{.validate_alpha}
\alias{.validate_group}
\alias{.validate_hyper}
\alias{.validate_foldid}
\title{Validation functions}
\usage{
.validate_na_action(na_action)

.validate_family(family, poisson = TRUE)

.validate_x(x, na_action)

.validate_y(y, family, n, na_action, names)

.validate_y_hat(y_hat, family, n)

.validate_primary(primary, p, names)

.validate_cor(cor, p, names)

.validate_alpha(alpha, init)

.validate_group(group, p, names)

.validate_hyper(hyper)

.validate_foldid(foldid, y, family)
}
\arguments{
\item{na_action}{character \code{"error"} to trigger an error
if any observation has a missing predictor or a missing response
or \code{"complete_cases"} to exclude observations
with a missing predictor or a missing response from model fitting
(while providing fitted values for these observations)}

\item{family}{character string \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, or \code{"cox"}}

\item{x}{\eqn{n_0 \times p} predictor matrix,
containing only numerical values (continuous, integer, or binary),
where \eqn{n_0} is the number of observations used for model training
and \eqn{p} is the number of predictors}

\item{y}{response vector of length \eqn{n_0},
containing numerical values (\code{family="gaussian"}),
integer values (\code{family="poisson"}),
binary values (\code{family="binomial"}),
or a survival object created with \code{survival::Surv()} (\code{family="cox"}),
where \eqn{n_0} is the number of observations used for model training}

\item{names}{character vector
of length \eqn{n} or \eqn{p}
for names of observations or predictors}

\item{y_hat}{\eqn{n_0}-dimensional vector of fitted values or probabilities}

\item{primary}{\eqn{p}-dimensional logical vector
indicating whether a predictor may be included in the final model
(\code{TRUE} for "primary predictors")
or must be excluded from the final model
(\code{FALSE} for "auxiliary predictors")}

\item{cor}{character string \code{"pearson"} (default),
\code{"spearman"},
or \code{"kendall"};
or a correlation matrix
(\eqn{p} rows, \eqn{p} columns,
entries between \eqn{-1} and \eqn{+1})}

\item{group}{group structure (multiple options):
\itemize{
\item \eqn{p}-dimensional vector of group indices
(in \eqn{\{1, \ldots, q\}}) or labels,
\item list with \eqn{q} slots containing the variable indices
(in \eqn{\{1, \ldots, p\}}) or labels,
\item \eqn{p \times p} matrix,
where the entry in the \eqn{j^{\text{th}}} row
and the \eqn{k^{\text{th}}} column
indicates whether information should be transferred
from the \eqn{j^{\text{th}}} to the \eqn{k^{\text{th}}} variable
(usually with a unit diagonal so that predictors retain the information
from their initial coefficient)
}}

\item{foldid}{\eqn{n_0}-dimensional vector containing the fold identifiers
(minimum \eqn{1}, maximum \code{nfolds})}
}
\value{
Return the first argument invisibly.
Throw an error for invalid arguments.
}
\description{
These functions validate the arguments
of the function \code{\link[=cv.corila]{cv.corila()}}, its helper functions, and its S3 methods.
They check whether the provided arguments satisfy expectations,
and return them in standardised forms
(e.g., as integers instead of integer-like numerics).
}
\details{
These functions are called by \code{\link[=cv.corila]{cv.corila()}},
its helper functions, and its S3 methods.
}
\keyword{internal}
